Contextual attention layer (2-D)

Contextual attention layer for generative image inpainting described in

Value

output tensor with the same shape as the input.

Details

Jiahui Yu, et al., Generative Image Inpainting with Contextual Attention, CVPR 2018.

available here:

    \code{https://arxiv.org/abs/1801.07892}

Usage

layer &lt;- ContextualAttentionLayer2D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )

Arguments

layer

A process object.

scale

feature scale. Default = 20

x

mask

input_shape

Details

$initialize instantiates a new class.

$build

$call main body.

$compute_output_shape computes the output shape.

Examples

x = ContextualAttentionLayer2D$new()
x$build()

#> NULL