Mean square error of a single image or between two images.

Mean square error of a single image or between two images.

MSE(x, y = NULL)

Arguments

x
input image.

y
input image.

Value

the mean squared error

Examples

library( ANTsR )

r16 <- antsImageRead( getANTsRData( 'r16' ) )
r85 <- antsImageRead( getANTsRData( 'r85' ) )
mseValue <- MSE( r16, r85 )