brainAge FMRIB

Estimate BrainAge from a T1-weighted MR image using the SCFN architecture and weights described here:

brainAgeFmrib(
  image,
  skipPreprocessing = FALSE,
  outputDirectory = NULL,
  verbose = FALSE
)

Arguments

image
input 3-D T1-weighted brain image.

skipPreprocessing
boolean dictating whether prescribed preprocessing is performed (brain extraction, bias correction, normalization to template).

outputDirectory
destination directory for storing the downloaded template and model weights. Since these can be resused, if is.null(outputDirectory), these data will be downloaded to the inst/extdata/ subfolder of the ANTsRNet package.

verbose
print progress.

Value

predicted age and binned confidence values

Details

https://github.com/ha-ha-ha-han/UKBiobank_deep_pretrain

and described in the following article:

https://doi.org/10.1101/2019.12.17.879346

Note that age prediction for this particular set of weights is in the range 42-82 years trained from the biobank data. The authors mention a separate data set from the PAC 2019 challenge where the range is 14-94 years but, as far as I can tell, those weights aren't available.

Examples

if (FALSE) {
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
brainAge <- brainAgeFmrib( image )
}