Brain extraction

Perform T1, FA, or bold brain extraction using a U-net architecture training data. "NoBrainer" is also possible where brain extraction uses U-net and FreeSurfer training data ported from the

brainExtraction(
  image,
  modality = c("t1", "t1v0", "t1nobrainer", "t1combined", "t2", "flair", "bold", "fa",
    "t1t2infant", "t1infant", "t2infant"),
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)

Arguments

image
input 3-D brain image (or list of images for multi-modal scenarios).

modality
image type. Options include:

"t1": T1-weighted MRI---ANTs-trained. Update from "t1v0"

"t1v0": T1-weighted MRI---ANTs-trained.

"t1nobrainer": T1-weighted MRI---FreeSurfer-trained: h/t Satra Ghosh and Jakub Kaczmarzyk.

"t1combined": Brian's combination of "t1" and "t1nobrainer". One can also specify "t1combinedX" where X is the morphological radius. X = 12 by default.

"flair": FLAIR MRI.

"t2": T2-w MRI.

"bold": 3-D BOLD MRI.

"fa": Fractional anisotropy.

"t1t2infant": Combined T1-w/T2-w infant MRI h/t Martin Styner.

"t1infant": T1-w infant MRI h/t Martin Styner.

"t2infant": T2-w infant MRI h/t Martin Styner.

antsxnetCacheDirectory
destination directory for storing the downloaded template and model weights. Since these can be resused, if is.null(antsxnetCacheDirectory), these data will be downloaded to the subdirectory ~/.keras/ANTsXNet/.

verbose
print progress.

Value

brain probability mask (ANTsR image)

Details

https://github.com/neuronets/nobrainer-models

Author

Tustison NJ

Examples

if (FALSE) {
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
probabilityMask <- brainExtraction( image, modality = "t1" )
}