Cerebellum tissue segmentation, Schmahmann parcellation, and thickness.

Perform cerebellum segmentation using a modification of the set of MaGET cerebellum atlases --- https://www.cobralab.ca/cerebellum-lobules.

cerebellumMorphology(
  t1,
  cerebellumMask = NULL,
  computeThicknessImage = FALSE,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)

Arguments

t1
raw or preprocessed 3-D T1-weighted whole head image.

cerebellumMask
Option for initialization. If not specified, the cerebellum ROI is determined using ANTsXNet brain_extraction followed by registration to a template.

computeThicknessImage
Compute KellyKapowski thickness image of the gray matter.

doPreprocessing
Perform N4 bias correction and spatiall normalize to template space.

antsxnetCacheDirectory
destination directory for storing the downloaded template and model weights. Since these can be ressed, if is.null(antsxnetCacheDirectory), these data will be downloaded to the inst/extdata/ subfolder of the ANTsRNet package.

verbose
print progress.

Value

List consisting of the multiple segmentation images and probability images for each label and foreground. Optional thickness image.

Details

The tissue labeling is as follows:

Label 1 :CSF

Label 2 :Gray matter

Label 3 :White matter

The parcellation labeling is as follows:

Label 1 :L_I_II

Label 2 :L_III

Label 3 :L_IV

Label 4 :L_V

Label 5 :L_VI

Label 6 :L_Crus_I

Label 7 :L_Crus_II

Label 8 :L_VIIB

Label 9 :L_VIIIA

Label 10 :L_VIIIB

Label 11 :L_IX

Label 12 :L_X

Label 101:R_I_II

Label 102:R_III

Label 103:R_IV

Label 104:R_V

Label 105:R_VI

Label 106:R_Crus_I

Label 107:R_Crus_II

Label 108:R_VIIB

Label 109:R_VIIIA

Label 110:R_VIIIB

Label 111:R_IX

Label 112:R_X

Author

Tustison NJ, Tustison MG

Examples

if (FALSE) {
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
cereb <- cerebellumMorphology( image )
}