2-D implementation of the AlexNet deep learning architecture.

Creates a keras model of the AlexNet deep learning architecture for image recognition based on the paper

createAlexNetModel2D(
  inputImageSize,
  numberOfClassificationLabels = 1000,
  numberOfDenseUnits = 4096,
  dropoutRate = 0,
  mode = c("classification", "regression"),
  batch_size = NULL
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfClassificationLabels
Number of segmentation labels.

numberOfDenseUnits
number of dense units.

dropoutRate
optional regularization parameter between [0, 1]. Default = 0.0.

mode
'classification' or 'regression'.

batch_size
batch size to pass to first layer

Value

an AlexNet keras model

Details

A. Krizhevsky, and I. Sutskever, and G. Hinton. ImageNet Classification with Deep Convolutional Neural Networks.

available here:

    http://papers.nips.cc/paper/4824-imagenet-classification-with-deep-convolutional-neural-networks.pdf

This particular implementation was influenced by the following python implementation:

    https://github.com/duggalrahul/AlexNet-Experiments-Keras/
    https://github.com/lunardog/convnets-keras/

Examples

library( ANTsRNet )
library( keras )

mnistData <- dataset_mnist()
numberOfLabels <- 10

# Extract a small subset for something that can run quickly

X_trainSmall <- mnistData$train$x[1:10,,]
X_trainSmall <- array( data = X_trainSmall, dim = c( dim( X_trainSmall ), 1 ) )
Y_trainSmall <- to_categorical( mnistData$train$y[1:10], numberOfLabels )

X_testSmall <- mnistData$test$x[1:10,,]
X_testSmall <- array( data = X_testSmall, dim = c( dim( X_testSmall ), 1 ) )
Y_testSmall <- to_categorical( mnistData$test$y[1:10], numberOfLabels )

# We add a dimension of 1 to specify the channel size

inputImageSize <- c( dim( X_trainSmall )[2:3], 1 )

model <- createAlexNetModel2D( inputImageSize = inputImageSize,
  numberOfClassificationLabels = numberOfLabels )

model %>% compile( loss = 'categorical_crossentropy',
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( 'categorical_crossentropy', 'accuracy' ) )
gc()

#>            used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells  2434108 130.0    4612587 246.4         NA   4612587  246.4
#> Vcells 31828700 242.9  122493408 934.6      16384 150387201 1147.4

rm(mnistData); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2431857 129.9    4612587 246.4         NA   4612587  246.4
#> Vcells 4349984  33.2   97994727 747.7      16384 150387201 1147.4

# Comment out the rest due to travis build constraints

# track <- model %>% fit( X_trainSmall, Y_trainSmall, verbose = 1,
#   epochs = 1, batch_size = 2, shuffle = TRUE, validation_split = 0.5 )

# Now test the model

# testingMetrics <- model %>% evaluate( X_testSmall, Y_testSmall )
# predictedData <- model %>% predict( X_testSmall, verbose = 1 )
rm(model); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2431918 129.9    4612587 246.4         NA   4612587  246.4
#> Vcells 4350051  33.2   78395782 598.2      16384 150387201 1147.4