Function for creating a 2-D symmetric convolutional autoencoder model.

Builds a convolutional autoencoder based on the specified array definining the number of units in the encoding branch. Ported to Keras R from the Keras python implementation here:

createConvolutionalAutoencoderModel2D(
  inputImageSize,
  numberOfFiltersPerLayer = c(32, 64, 128, 10),
  convolutionKernelSize = c(5, 5),
  deconvolutionKernelSize = c(5, 5)
)

Arguments

inputImageSize
vector definining spatial dimensions + channels

numberOfFiltersPerLayer
vector defining the number of convolutional filters in the encoding branch per layer

convolutionKernelSize
kernel size fo the convolutional filters

deconvolutionKernelSize
kernel size fo the convolutional transpose filters

Value

two models: the convolutional encoder and convolutional auto-encoder

Details

https://github.com/XifengGuo/DCEC

Examples

library( ANTsRNet )
library( keras )

ae <- createConvolutionalAutoencoderModel2D( c( 32, 32, 1 ) )