3-D implementation of the denoising autoencoder image super resolution architecture.

3-D implementation of the denoising autoencoder image super resolution architecture.

createDenoisingAutoEncoderSuperResolutionModel3D(
  inputImageSize,
  convolutionKernelSizes = list(c(3, 3, 3), c(5, 5, 5)),
  numberOfEncodingLayers = 2,
  numberOfFilters = 64
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSizes
a 2-element list of 3-D vectors specifying the kernel size at each convolution layer. The first element is the kernel size of the encoding layers and the 2nd element is the kernel size of the final convolution layer.

numberOfEncodingLayers
the number of encoding layers.

numberOfFilters
the number of filters for each encoding layer.

Value

a keras model for image super resolution

Examples

createDenoisingAutoEncoderSuperResolutionModel3D(c( 28, 28, 28, 1 ))

#> Model
#> Model: "model_17"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_14 (InputLayer)     [(None, 28, 28, 2 0                                   
#> ________________________________________________________________________________
#> conv3d_40 (Conv3D)        (None, 28, 28, 28 1792     input_14[0][0]             
#> ________________________________________________________________________________
#> conv3d_41 (Conv3D)        (None, 28, 28, 28 110656   conv3d_40[0][0]            
#> ________________________________________________________________________________
#> conv3d_transpose_23 (Conv (None, 28, 28, 28 110656   conv3d_41[0][0]            
#> ________________________________________________________________________________
#> add_32 (Add)              (None, 28, 28, 28 0        conv3d_41[0][0]            
#>                                                      conv3d_transpose_23[0][0]  
#> ________________________________________________________________________________
#> conv3d_transpose_24 (Conv (None, 28, 28, 28 110656   add_32[0][0]               
#> ________________________________________________________________________________
#> add_33 (Add)              (None, 28, 28, 28 0        conv3d_40[0][0]            
#>                                                      conv3d_transpose_24[0][0]  
#> ________________________________________________________________________________
#> conv3d_42 (Conv3D)        (None, 28, 28, 28 8001     add_33[0][0]               
#> ================================================================================
#> Total params: 341,761
#> Trainable params: 341,761
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> 
#> 

gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2481566 132.6    3758494 200.8         NA   3758494 200.8
#> Vcells 5299031  40.5   40399536 308.3      16384 128142950 977.7