2-D implementation of the dense U-net deep learning architecture.

Creates a keras model of the dense U-net deep learning architecture for image segmentation

createDenseUnetModel2D(
  inputImageSize,
  numberOfOutputs = 1L,
  numberOfLayersPerDenseBlock = c(6, 12, 36, 24),
  growthRate = 48,
  initialNumberOfFilters = 96,
  reductionRate = 0,
  depth = 7,
  dropoutRate = 0,
  weightDecay = 0.0001,
  mode = c("classification", "regression")
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue).

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

numberOfLayersPerDenseBlock
number of dense blocks per layer.

growthRate
number of filters to add for each dense block layer (default = 48).

initialNumberOfFilters
number of filters at the beginning (default = 96).

reductionRate
reduction factor of transition blocks

depth
number of layers---must be equal to 3 * N + 4 where N is an integer (default = 7).

dropoutRate
drop out layer rate (default = 0.2).

weightDecay
weight decay (default = 1e-4).

mode
A switch to determine the activation function to use. If classification, then sigmoid or softmax depending on the numberOfOutputs, and if mode = "regression" then function is linear

Value

an DenseUnet keras model

Details

X. Li, H. Chen, X. Qi, Q. Dou, C.-W. Fu, P.-A. Heng. H-DenseUNet: Hybrid Densely Connected UNet for Liver and Tumor Segmentation from CT Volumes

available here:

    https://arxiv.org/pdf/1709.07330.pdf

with the author's implementation available at:

    https://github.com/xmengli999/H-DenseUNet

Examples

library( ANTsR )
library( ANTsRNet )
library( keras )

imageIDs <- c( "r16", "r27", "r30", "r62", "r64", "r85" )
trainingBatchSize <- length( imageIDs )

# Perform simple 3-tissue segmentation.

segmentationLabels <- c( 1, 2, 3 )
numberOfLabels <- length( segmentationLabels )
initialization <- paste0( 'KMeans[', numberOfLabels, ']' )

domainImage <- antsImageRead( getANTsRData( imageIDs[1] ) )

X_train <- array( data = NA, dim = c( trainingBatchSize, dim( domainImage ), 1 ) )
Y_train <- array( data = NA, dim = c( trainingBatchSize, dim( domainImage ) ) )

images <- list()
segmentations <- list()

for( i in seq_len( trainingBatchSize ) )
  {
  cat( "Processing image", imageIDs[i], "\n" )
  image <- antsImageRead( getANTsRData( imageIDs[i] ) )
  mask <- getMask( image )
  segmentation <- atropos( image, mask, initialization )$segmentation

  X_train[i,,, 1] <- as.array( image )
  Y_train[i,,] <- as.array( segmentation )
  }

#> Processing image r16 
#> Processing image r27 
#> Processing image r30 
#> Processing image r62 
#> Processing image r64 
#> Processing image r85 

Y_train <- encodeUnet( Y_train, segmentationLabels )

# Perform a simple normalization

X_train <- ( X_train - mean( X_train ) ) / sd( X_train )

# Create the model

model <- createDenseUnetModel2D( c( dim( domainImage ), 1L ),
  numberOfOutputs = numberOfLabels )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_multilabel_dice_coefficient,
    metric_categorical_crossentropy ) )

# Comment out the rest due to travis build constraints

# Fit the model
# \donttest{
checkpoint_file = tempfile(fileext = ".h5")
track <- model %>% fit( X_train, Y_train,
              epochs = 5, batch_size = 4, verbose = 1, shuffle = TRUE,
              callbacks = list(
                callback_model_checkpoint( checkpoint_file,
                    monitor = 'val_loss', save_best_only = TRUE ),
                callback_reduce_lr_on_plateau( monitor = "val_loss", factor = 0.1 )
              ),
              validation_split = 0.2 )

#> Error in py_call_impl(callable, dots$args, dots$keywords): RuntimeError: Evaluation error: ValueError: Dimension must be 4 but is 5 for 'loss/conv2d_259_loss/transpose' (op: 'Transpose') with input shapes: [4,256,256,3], [5].
#> 
#> Detailed traceback: 
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/backend.py", line 2768, in permute_dimensions
#>     return array_ops.transpose(x, perm=pattern)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/ops/array_ops.py", line 1870, in transpose
#>     ret = transpose_fn(a, perm, name=name)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/ops/gen_array_ops.py", line 11455, in transpose
#>     "Transpose", x=x, perm=perm, name=name)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/op_def_library.py", line 793, in _apply_op_helper
#>     op_def=op_def)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/func_graph.py", line 548, in create_op
#>     compute_device)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/ops.py", line 3429, in _create_op_internal
#>     op_def=op_def)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/ops.py", line 1773, in __init__
#>     control_input_ops)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/ops.py", line 1613, in _create_c_op
#>     raise ValueError(str(e))
#> .
#> 
#> Detailed traceback: 
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training.py", line 728, in fit
#>     use_multiprocessing=use_multiprocessing)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_v2.py", line 324, in fit
#>     total_epochs=epochs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_v2.py", line 123, in run_one_epoch
#>     batch_outs = execution_function(iterator)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_v2_utils.py", line 86, in execution_function
#>     distributed_function(input_fn))
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/def_function.py", line 457, in __call__
#>     result = self._call(*args, **kwds)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/def_function.py", line 503, in _call
#>     self._initialize(args, kwds, add_initializers_to=initializer_map)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/def_function.py", line 408, in _initialize
#>     *args, **kwds))
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/function.py", line 1848, in _get_concrete_function_internal_garbage_collected
#>     graph_function, _, _ = self._maybe_define_function(args, kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/function.py", line 2150, in _maybe_define_function
#>     graph_function = self._create_graph_function(args, kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/function.py", line 2041, in _create_graph_function
#>     capture_by_value=self._capture_by_value),
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/framework/func_graph.py", line 915, in func_graph_from_py_func
#>     func_outputs = python_func(*func_args, **func_kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/eager/def_function.py", line 358, in wrapped_fn
#>     return weak_wrapped_fn().__wrapped__(*args, **kwds)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_v2_utils.py", line 73, in distributed_function
#>     per_replica_function, args=(model, x, y, sample_weights))
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/distribute/distribute_lib.py", line 760, in experimental_run_v2
#>     return self._extended.call_for_each_replica(fn, args=args, kwargs=kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/distribute/distribute_lib.py", line 1787, in call_for_each_replica
#>     return self._call_for_each_replica(fn, args, kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/distribute/distribute_lib.py", line 2132, in _call_for_each_replica
#>     return fn(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/autograph/impl/api.py", line 292, in wrapper
#>     return func(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_v2_utils.py", line 264, in train_on_batch
#>     output_loss_metrics=model._output_loss_metrics)
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_eager.py", line 311, in train_on_batch
#>     output_loss_metrics=output_loss_metrics))
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_eager.py", line 252, in _process_single_batch
#>     training=training))
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/engine/training_eager.py", line 166, in _model_loss
#>     per_sample_losses = loss_fn.call(targets[i], outs[i])
#>   File "/Users/ntustison/anaconda3/envs/r-tensorflow/lib/python3.6/site-packages/tensorflow_core/python/keras/losses.py", line 221, in call
#>     return self.fn(y_true, y_pred, **self._fn_kwargs)
#>   File "/Library/Frameworks/R.framework/Versions/3.6/Resources/library/reticulate/python/rpytools/call.py", line 21, in python_function
#>     raise RuntimeError(res[kErrorKey])

# }

# Save the model and/or save the model weights

# save_model_hdf5( model, filepath = 'unetModel.h5' )
# save_model_weights_hdf5( unetModel, filepath = 'unetModelWeights.h5' ) )
rm(model); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2491642 133.1    4110641 219.6         NA   4110641 219.6
#> Vcells 6057639  46.3   32099607 245.0      16384 130717590 997.3