2-D implementation of the EDSR super resolution architecture.

Creates a keras model of the expanded image super resolution deep learning framework based on EDSR.

createEnhancedDeepSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSize = c(3, 3),
  numberOfFilters = 256,
  numberOfResidualBlocks = 32,
  scale = 2,
  numberOfLossFunctions = 1,
  numberOfOutputChannels = 1,
  doBatchNormalization = FALSE,
  interpolation = c("bilinear", "nearest", "conv")
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSize
a vector specifying the kernel size for convolution.

numberOfFilters
the number of filters for each encoding layer.

numberOfResidualBlocks
the number of residual blocks.

scale
the upsampling amount, 2, 4 or 8

numberOfLossFunctions
the number of data targets, e.g. 2 for 2 targets

numberOfOutputChannels
the number of ouput channels

doBatchNormalization
boolean for include BN in the residual blocks

interpolation
nearest, bilinear or conv for upscaling block

Value

a keras model for EDSR image super resolution

Examples

model = createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ))
rm(model); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2473586 132.2    4612730 246.4         NA   4612730  246.4
#> Vcells 4392045  33.6   20550985 156.8      16384 150387071 1147.4

model = createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ),
doBatchNormalization = TRUE,
interpolation = "conv", scale = 4)
rm(model); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2474359 132.2    4612730 246.4         NA   4612730  246.4
#> Vcells 4393345  33.6   16440788 125.5      16384 150387071 1147.4

model = createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ),
doBatchNormalization = TRUE,
numberOfLossFunctions = 2,
interpolation = "conv", scale = 8)
rm(model); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2474016 132.2    4612730 246.4         NA   4612730  246.4
#> Vcells 4391790  33.6   16440788 125.5      16384 150387071 1147.4