3-D u-net architecture for hypothalamus segmentation

Implementation of the U-net architecture for hypothalamus segmentation described in

createHypothalamusUnetModel3D(inputImageSize)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions only (number of channels is 1).

Value

a u-net keras model

Details

https://pubmed.ncbi.nlm.nih.gov/32853816/

and ported from the original implementation:

   https://github.com/BBillot/hypothalamus_seg

The network has is characterized by the following parameters:

3 resolution levels: 24 ---> 48 ---> 96 filters

convolution: kernel size: c(3, 3, 3), activation: 'elu'

pool size: c(2, 2, 2)

Author

Tustison NJ

Examples

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )

model <- createHypothalamusUnetModel3d( c( 160, 160, 160 ) )

#> Error in createHypothalamusUnetModel3d(c(160, 160, 160)): could not find function "createHypothalamusUnetModel3d"

print( model )

#> Error in print(model): object 'model' not found