2-D implementation of the image super resolution deep learning architecture.

Creates a keras model of the image super resolution deep learning framework. based on the paper available here:

createImageSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSizes = list(c(9, 9), c(1, 1), c(5, 5)),
  numberOfFilters = c(64, 32)
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSizes
a list of 2-D vectors specifying the kernel size at each convolution layer. Default values are the same as given in the original paper. The length of kernel size vectors must be 1 greater than the vector length of the number of filters.

numberOfFilters
a vector containing the number of filters for each convolutional layer. Default values are the same as given in the original paper.

Value

a keras model for image super resolution

Details

    \url{https://arxiv.org/pdf/1501.00092}

This particular implementation is based on the following python implementation:

    \url{https://github.com/titu1994/Image-Super-Resolution}

Examples

createImageSuperResolutionModel2D(c( 100, 100, 1 ))

#> Model
#> Model: "model_29"
#> ________________________________________________________________________________
#> Layer (type)                        Output Shape                    Param #     
#> ================================================================================
#> input_26 (InputLayer)               [(None, 100, 100, 1)]           0           
#> ________________________________________________________________________________
#> conv2d_748 (Conv2D)                 (None, 100, 100, 64)            5248        
#> ________________________________________________________________________________
#> conv2d_749 (Conv2D)                 (None, 100, 100, 32)            2080        
#> ________________________________________________________________________________
#> conv2d_750 (Conv2D)                 (None, 100, 100, 1)             801         
#> ================================================================================
#> Total params: 8,129
#> Trainable params: 8,129
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> 
#> 

gc()

#>           used (Mb) gc trigger  (Mb) limit (Mb)  max used   (Mb)
#> Ncells 2471241  132    4427115 236.5         NA   4427115  236.5
#> Vcells 6020078   46   43383560 331.0      16384 150674493 1149.6