3-D implementation of the "NoBrainer" U-net architecture

Creates a keras model implementation of the u-net architecture avaialable here:

createNoBrainerUnetModel3D(inputImageSize, numberOfOutputs = 1,
  numberOfFiltersAtBaseLayer = 16, convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2))

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue).

numberOfOutputs
Number of segmentation labels.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer. Default = 16.

convolutionKernelSize
3-d vector defining the kernel size during the encoding path.

deconvolutionKernelSize
3-d vector defining the kernel size during the decoding.

Value

a u-net keras model

Details

https://github.com/neuronets/nobrainer/blob/master/nobrainer/models/unet.py

Examples

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )
library( keras )

model <- createNoBrainerUnetModel3D( c( 128, 128, 128, 1 ) )
url <- "https://github.com/neuronets/nobrainer-models/releases/download/0.1/brain-extraction-unet-128iso-weights.h5"
weightsFile <- "nobrainerWeights.h5"
download.file( url, weightsFile )
model$load_weights( weightsFile )

print( model )

#> Model
#> Model: "model_14"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_11 (InputLayer)     (None, 128, 128,  0                                   
#> ________________________________________________________________________________
#> conv3d_64 (Conv3D)        (None, 128, 128,  448      input_11[0][0]             
#> ________________________________________________________________________________
#> re_lu_1 (ReLU)            (None, 128, 128,  0        conv3d_64[0][0]            
#> ________________________________________________________________________________
#> conv3d_65 (Conv3D)        (None, 128, 128,  13856    re_lu_1[0][0]              
#> ________________________________________________________________________________
#> re_lu_2 (ReLU)            (None, 128, 128,  0        conv3d_65[0][0]            
#> ________________________________________________________________________________
#> max_pooling3d_2 (MaxPooli (None, 64, 64, 64 0        re_lu_2[0][0]              
#> ________________________________________________________________________________
#> conv3d_66 (Conv3D)        (None, 64, 64, 64 27680    max_pooling3d_2[0][0]      
#> ________________________________________________________________________________
#> re_lu_3 (ReLU)            (None, 64, 64, 64 0        conv3d_66[0][0]            
#> ________________________________________________________________________________
#> conv3d_67 (Conv3D)        (None, 64, 64, 64 55360    re_lu_3[0][0]              
#> ________________________________________________________________________________
#> re_lu_4 (ReLU)            (None, 64, 64, 64 0        conv3d_67[0][0]            
#> ________________________________________________________________________________
#> max_pooling3d_3 (MaxPooli (None, 32, 32, 32 0        re_lu_4[0][0]              
#> ________________________________________________________________________________
#> conv3d_68 (Conv3D)        (None, 32, 32, 32 110656   max_pooling3d_3[0][0]      
#> ________________________________________________________________________________
#> re_lu_5 (ReLU)            (None, 32, 32, 32 0        conv3d_68[0][0]            
#> ________________________________________________________________________________
#> conv3d_69 (Conv3D)        (None, 32, 32, 32 221312   re_lu_5[0][0]              
#> ________________________________________________________________________________
#> re_lu_6 (ReLU)            (None, 32, 32, 32 0        conv3d_69[0][0]            
#> ________________________________________________________________________________
#> max_pooling3d_4 (MaxPooli (None, 16, 16, 16 0        re_lu_6[0][0]              
#> ________________________________________________________________________________
#> conv3d_70 (Conv3D)        (None, 16, 16, 16 442496   max_pooling3d_4[0][0]      
#> ________________________________________________________________________________
#> re_lu_7 (ReLU)            (None, 16, 16, 16 0        conv3d_70[0][0]            
#> ________________________________________________________________________________
#> conv3d_71 (Conv3D)        (None, 16, 16, 16 884992   re_lu_7[0][0]              
#> ________________________________________________________________________________
#> re_lu_8 (ReLU)            (None, 16, 16, 16 0        conv3d_71[0][0]            
#> ________________________________________________________________________________
#> conv3d_transpose_1 (Conv3 (None, 32, 32, 32 524544   re_lu_8[0][0]              
#> ________________________________________________________________________________
#> concatenate_131 (Concaten (None, 32, 32, 32 0        re_lu_6[0][0]              
#>                                                      conv3d_transpose_1[0][0]   
#> ________________________________________________________________________________
#> conv3d_72 (Conv3D)        (None, 32, 32, 32 1327232  concatenate_131[0][0]      
#> ________________________________________________________________________________
#> re_lu_9 (ReLU)            (None, 32, 32, 32 0        conv3d_72[0][0]            
#> ________________________________________________________________________________
#> conv3d_73 (Conv3D)        (None, 32, 32, 32 442496   re_lu_9[0][0]              
#> ________________________________________________________________________________
#> re_lu_10 (ReLU)           (None, 32, 32, 32 0        conv3d_73[0][0]            
#> ________________________________________________________________________________
#> conv3d_transpose_2 (Conv3 (None, 64, 64, 64 131200   re_lu_10[0][0]             
#> ________________________________________________________________________________
#> concatenate_132 (Concaten (None, 64, 64, 64 0        re_lu_4[0][0]              
#>                                                      conv3d_transpose_2[0][0]   
#> ________________________________________________________________________________
#> conv3d_74 (Conv3D)        (None, 64, 64, 64 331840   concatenate_132[0][0]      
#> ________________________________________________________________________________
#> re_lu_11 (ReLU)           (None, 64, 64, 64 0        conv3d_74[0][0]            
#> ________________________________________________________________________________
#> conv3d_75 (Conv3D)        (None, 64, 64, 64 110656   re_lu_11[0][0]             
#> ________________________________________________________________________________
#> re_lu_12 (ReLU)           (None, 64, 64, 64 0        conv3d_75[0][0]            
#> ________________________________________________________________________________
#> conv3d_transpose_3 (Conv3 (None, 128, 128,  32832    re_lu_12[0][0]             
#> ________________________________________________________________________________
#> concatenate_133 (Concaten (None, 128, 128,  0        re_lu_2[0][0]              
#>                                                      conv3d_transpose_3[0][0]   
#> ________________________________________________________________________________
#> conv3d_76 (Conv3D)        (None, 128, 128,  82976    concatenate_133[0][0]      
#> ________________________________________________________________________________
#> re_lu_13 (ReLU)           (None, 128, 128,  0        conv3d_76[0][0]            
#> ________________________________________________________________________________
#> conv3d_77 (Conv3D)        (None, 128, 128,  27680    re_lu_13[0][0]             
#> ________________________________________________________________________________
#> re_lu_14 (ReLU)           (None, 128, 128,  0        conv3d_77[0][0]            
#> ________________________________________________________________________________
#> conv3d_78 (Conv3D)        (None, 128, 128,  33       re_lu_14[0][0]             
#> ================================================================================
#> Total params: 4,768,289
#> Trainable params: 4,768,289
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> 
#>