2-D implementation of the ResNet image super resolution architecture.

Creates a keras model of the expanded image super resolution deep learning framework based on the following python implementation:

createResNetSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSize = c(3, 3),
  numberOfFilters = 64,
  numberOfResidualBlocks = 5,
  numberOfResNetBlocks = 1
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSize
a vector specifying the kernel size for convolution.

numberOfFilters
the number of filters for each encoding layer.

numberOfResidualBlocks
the number of residual blocks.

numberOfResNetBlocks
the number of resNet blocks. Each block will double the upsampling amount.

Value

a keras model for ResNet image super resolution

Details

    \url{https://github.com/titu1994/Image-Super-Resolution}

Examples

createResNetSuperResolutionModel2D(c(256, 256, 3))

#> Model
#> Model: "model_33"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_30 (InputLayer)     [(None, 256, 256, 0                                   
#> ________________________________________________________________________________
#> conv2d_804 (Conv2D)       (None, 256, 256,  1792     input_30[0][0]             
#> ________________________________________________________________________________
#> conv2d_805 (Conv2D)       (None, 256, 256,  36928    conv2d_804[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_632 ( (None, 256, 256,  256      conv2d_805[0][0]           
#> ________________________________________________________________________________
#> activation_513 (Activatio (None, 256, 256,  0        batch_normalization_632[0][
#> ________________________________________________________________________________
#> conv2d_806 (Conv2D)       (None, 256, 256,  36928    activation_513[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_633 ( (None, 256, 256,  256      conv2d_806[0][0]           
#> ________________________________________________________________________________
#> add_155 (Add)             (None, 256, 256,  0        conv2d_804[0][0]           
#>                                                      batch_normalization_633[0][
#> ________________________________________________________________________________
#> conv2d_807 (Conv2D)       (None, 256, 256,  36928    add_155[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_634 ( (None, 256, 256,  256      conv2d_807[0][0]           
#> ________________________________________________________________________________
#> activation_514 (Activatio (None, 256, 256,  0        batch_normalization_634[0][
#> ________________________________________________________________________________
#> conv2d_808 (Conv2D)       (None, 256, 256,  36928    activation_514[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_635 ( (None, 256, 256,  256      conv2d_808[0][0]           
#> ________________________________________________________________________________
#> add_156 (Add)             (None, 256, 256,  0        add_155[0][0]              
#>                                                      batch_normalization_635[0][
#> ________________________________________________________________________________
#> conv2d_809 (Conv2D)       (None, 256, 256,  36928    add_156[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_636 ( (None, 256, 256,  256      conv2d_809[0][0]           
#> ________________________________________________________________________________
#> activation_515 (Activatio (None, 256, 256,  0        batch_normalization_636[0][
#> ________________________________________________________________________________
#> conv2d_810 (Conv2D)       (None, 256, 256,  36928    activation_515[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_637 ( (None, 256, 256,  256      conv2d_810[0][0]           
#> ________________________________________________________________________________
#> add_157 (Add)             (None, 256, 256,  0        add_156[0][0]              
#>                                                      batch_normalization_637[0][
#> ________________________________________________________________________________
#> conv2d_811 (Conv2D)       (None, 256, 256,  36928    add_157[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_638 ( (None, 256, 256,  256      conv2d_811[0][0]           
#> ________________________________________________________________________________
#> activation_516 (Activatio (None, 256, 256,  0        batch_normalization_638[0][
#> ________________________________________________________________________________
#> conv2d_812 (Conv2D)       (None, 256, 256,  36928    activation_516[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_639 ( (None, 256, 256,  256      conv2d_812[0][0]           
#> ________________________________________________________________________________
#> add_158 (Add)             (None, 256, 256,  0        add_157[0][0]              
#>                                                      batch_normalization_639[0][
#> ________________________________________________________________________________
#> conv2d_813 (Conv2D)       (None, 256, 256,  36928    add_158[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_640 ( (None, 256, 256,  256      conv2d_813[0][0]           
#> ________________________________________________________________________________
#> activation_517 (Activatio (None, 256, 256,  0        batch_normalization_640[0][
#> ________________________________________________________________________________
#> conv2d_814 (Conv2D)       (None, 256, 256,  36928    activation_517[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_641 ( (None, 256, 256,  256      conv2d_814[0][0]           
#> ________________________________________________________________________________
#> add_159 (Add)             (None, 256, 256,  0        add_158[0][0]              
#>                                                      batch_normalization_641[0][
#> ________________________________________________________________________________
#> conv2d_815 (Conv2D)       (None, 256, 256,  36928    add_159[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_642 ( (None, 256, 256,  256      conv2d_815[0][0]           
#> ________________________________________________________________________________
#> activation_518 (Activatio (None, 256, 256,  0        batch_normalization_642[0][
#> ________________________________________________________________________________
#> conv2d_816 (Conv2D)       (None, 256, 256,  36928    activation_518[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_643 ( (None, 256, 256,  256      conv2d_816[0][0]           
#> ________________________________________________________________________________
#> add_160 (Add)             (None, 256, 256,  0        add_159[0][0]              
#>                                                      batch_normalization_643[0][
#> ________________________________________________________________________________
#> add_161 (Add)             (None, 256, 256,  0        add_160[0][0]              
#>                                                      conv2d_804[0][0]           
#> ________________________________________________________________________________
#> up_sampling2d_10 (UpSampl (None, 512, 512,  0        add_161[0][0]              
#> ________________________________________________________________________________
#> conv2d_817 (Conv2D)       (None, 512, 512,  36928    up_sampling2d_10[0][0]     
#> ________________________________________________________________________________
#> conv2d_818 (Conv2D)       (None, 512, 512,  1731     conv2d_817[0][0]           
#> ================================================================================
#> Total params: 486,659
#> Trainable params: 485,123
#> Non-trainable params: 1,536
#> ________________________________________________________________________________
#> 
#>