3-D implementation of the ResNet image super resolution architecture.

Creates a keras model of the expanded image super resolution deep learning framework based on the following python implementation:

createResNetSuperResolutionModel3D(
  inputImageSize,
  convolutionKernelSize = c(3, 3, 3),
  numberOfFilters = 64,
  numberOfResidualBlocks = 5,
  numberOfResNetBlocks = 1
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSize
a vector specifying the kernel size for convolution.

numberOfFilters
the number of filters for each encoding layer.

numberOfResidualBlocks
the number of residual blocks.

numberOfResNetBlocks
the number of resNet blocks. Each block will double the upsampling amount.

Value

a keras model for ResNet image super resolution

Details

    \url{https://github.com/titu1994/Image-Super-Resolution}

Examples

inputImageSize = c(256, 256, 30, 1)
createResNetSuperResolutionModel3D(inputImageSize)

#> Model
#> Model: "model_24"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_30 (InputLayer)     [(None, 256, 256, 0                                   
#> ________________________________________________________________________________
#> conv3d_85 (Conv3D)        (None, 256, 256,  1792     input_30[0][0]             
#> ________________________________________________________________________________
#> conv3d_86 (Conv3D)        (None, 256, 256,  110656   conv3d_85[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_655 ( (None, 256, 256,  256      conv3d_86[0][0]            
#> ________________________________________________________________________________
#> activation_516 (Activatio (None, 256, 256,  0        batch_normalization_655[0][
#> ________________________________________________________________________________
#> conv3d_87 (Conv3D)        (None, 256, 256,  110656   activation_516[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_656 ( (None, 256, 256,  256      conv3d_87[0][0]            
#> ________________________________________________________________________________
#> add_63 (Add)              (None, 256, 256,  0        conv3d_85[0][0]            
#>                                                      batch_normalization_656[0][
#> ________________________________________________________________________________
#> conv3d_88 (Conv3D)        (None, 256, 256,  110656   add_63[0][0]               
#> ________________________________________________________________________________
#> batch_normalization_657 ( (None, 256, 256,  256      conv3d_88[0][0]            
#> ________________________________________________________________________________
#> activation_517 (Activatio (None, 256, 256,  0        batch_normalization_657[0][
#> ________________________________________________________________________________
#> conv3d_89 (Conv3D)        (None, 256, 256,  110656   activation_517[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_658 ( (None, 256, 256,  256      conv3d_89[0][0]            
#> ________________________________________________________________________________
#> add_64 (Add)              (None, 256, 256,  0        add_63[0][0]               
#>                                                      batch_normalization_658[0][
#> ________________________________________________________________________________
#> conv3d_90 (Conv3D)        (None, 256, 256,  110656   add_64[0][0]               
#> ________________________________________________________________________________
#> batch_normalization_659 ( (None, 256, 256,  256      conv3d_90[0][0]            
#> ________________________________________________________________________________
#> activation_518 (Activatio (None, 256, 256,  0        batch_normalization_659[0][
#> ________________________________________________________________________________
#> conv3d_91 (Conv3D)        (None, 256, 256,  110656   activation_518[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_660 ( (None, 256, 256,  256      conv3d_91[0][0]            
#> ________________________________________________________________________________
#> add_65 (Add)              (None, 256, 256,  0        add_64[0][0]               
#>                                                      batch_normalization_660[0][
#> ________________________________________________________________________________
#> conv3d_92 (Conv3D)        (None, 256, 256,  110656   add_65[0][0]               
#> ________________________________________________________________________________
#> batch_normalization_661 ( (None, 256, 256,  256      conv3d_92[0][0]            
#> ________________________________________________________________________________
#> activation_519 (Activatio (None, 256, 256,  0        batch_normalization_661[0][
#> ________________________________________________________________________________
#> conv3d_93 (Conv3D)        (None, 256, 256,  110656   activation_519[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_662 ( (None, 256, 256,  256      conv3d_93[0][0]            
#> ________________________________________________________________________________
#> add_66 (Add)              (None, 256, 256,  0        add_65[0][0]               
#>                                                      batch_normalization_662[0][
#> ________________________________________________________________________________
#> conv3d_94 (Conv3D)        (None, 256, 256,  110656   add_66[0][0]               
#> ________________________________________________________________________________
#> batch_normalization_663 ( (None, 256, 256,  256      conv3d_94[0][0]            
#> ________________________________________________________________________________
#> activation_520 (Activatio (None, 256, 256,  0        batch_normalization_663[0][
#> ________________________________________________________________________________
#> conv3d_95 (Conv3D)        (None, 256, 256,  110656   activation_520[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_664 ( (None, 256, 256,  256      conv3d_95[0][0]            
#> ________________________________________________________________________________
#> add_67 (Add)              (None, 256, 256,  0        add_66[0][0]               
#>                                                      batch_normalization_664[0][
#> ________________________________________________________________________________
#> conv3d_96 (Conv3D)        (None, 256, 256,  110656   add_67[0][0]               
#> ________________________________________________________________________________
#> batch_normalization_665 ( (None, 256, 256,  256      conv3d_96[0][0]            
#> ________________________________________________________________________________
#> activation_521 (Activatio (None, 256, 256,  0        batch_normalization_665[0][
#> ________________________________________________________________________________
#> conv3d_97 (Conv3D)        (None, 256, 256,  110656   activation_521[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_666 ( (None, 256, 256,  256      conv3d_97[0][0]            
#> ________________________________________________________________________________
#> add_68 (Add)              (None, 256, 256,  0        add_67[0][0]               
#>                                                      batch_normalization_666[0][
#> ________________________________________________________________________________
#> add_69 (Add)              (None, 256, 256,  0        add_68[0][0]               
#>                                                      conv3d_85[0][0]            
#> ________________________________________________________________________________
#> up_sampling3d_4 (UpSampli (None, 512, 512,  0        add_69[0][0]               
#> ________________________________________________________________________________
#> conv3d_98 (Conv3D)        (None, 512, 512,  110656   up_sampling3d_4[0][0]      
#> ________________________________________________________________________________
#> conv3d_99 (Conv3D)        (None, 512, 512,  1729     conv3d_98[0][0]            
#> ================================================================================
#> Total params: 1,445,121
#> Trainable params: 1,443,585
#> Non-trainable params: 1,536
#> ________________________________________________________________________________
#> 
#>