3-D implementation of the Resnet + U-net deep learning architecture.

Creates a keras model of the U-net + ResNet deep learning architecture for image segmentation and regression with the paper available here:

createResUnetModel3D(
  inputImageSize,
  numberOfOutputs = 1,
  numberOfFiltersAtBaseLayer = 32,
  bottleNeckBlockDepthSchedule = c(3, 4),
  convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2),
  dropoutRate = 0,
  weightDecay = 1e-04,
  mode = "classification"
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer.

bottleNeckBlockDepthSchedule
vector that provides the encoding layer schedule for the number of bottleneck blocks per long skip connection.

convolutionKernelSize
2-d vector defining the kernel size during the encoding path

deconvolutionKernelSize
2-d vector defining the kernel size during the decoding

dropoutRate
float between 0 and 1 to use between dense layers.

weightDecay
weighting parameter for L2 regularization of the kernel weights of the convolution layers. Default = 0.0.

mode
'classification' or 'regression'. Default = 'classification'.

Value

a res/u-net keras model

Details

    \url{https://arxiv.org/abs/1608.04117}

This particular implementation was ported from the following python implementation:

    \url{https://github.com/veugene/fcn_maker/}

Examples

library( ANTsRNet )
library( keras )

model <- createResUnetModel3D( c( 64, 64, 64, 1 ) )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_multilabel_dice_coefficient,
    metric_categorical_crossentropy ) )

print( model )

#> Model
#> Model: "model_26"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_32 (InputLayer)     [(None, 64, 64, 6 0                                   
#> ________________________________________________________________________________
#> conv3d_100 (Conv3D)       (None, 64, 64, 64 896      input_32[0][0]             
#> ________________________________________________________________________________
#> batch_normalization_705 ( (None, 64, 64, 64 128      conv3d_100[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_38 (Thr (None, 64, 64, 64 0        batch_normalization_705[0][
#> ________________________________________________________________________________
#> max_pooling3d_6 (MaxPooli (None, 32, 32, 32 0        thresholded_re_lu_38[0][0] 
#> ________________________________________________________________________________
#> conv3d_102 (Conv3D)       (None, 32, 32, 32 1056     conv3d_100[0][0]           
#> ________________________________________________________________________________
#> conv3d_101 (Conv3D)       (None, 32, 32, 32 27680    max_pooling3d_6[0][0]      
#> ________________________________________________________________________________
#> add_91 (Add)              (None, 32, 32, 32 0        conv3d_102[0][0]           
#>                                                      conv3d_101[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_706 ( (None, 32, 32, 32 128      add_91[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_39 (Thr (None, 32, 32, 32 0        batch_normalization_706[0][
#> ________________________________________________________________________________
#> conv3d_103 (Conv3D)       (None, 16, 16, 16 1056     thresholded_re_lu_39[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_707 ( (None, 16, 16, 16 128      conv3d_103[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_40 (Thr (None, 16, 16, 16 0        batch_normalization_707[0][
#> ________________________________________________________________________________
#> conv3d_104 (Conv3D)       (None, 16, 16, 16 1056     thresholded_re_lu_40[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_708 ( (None, 16, 16, 16 128      conv3d_104[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_41 (Thr (None, 16, 16, 16 0        batch_normalization_708[0][
#> ________________________________________________________________________________
#> conv3d_106 (Conv3D)       (None, 16, 16, 16 4224     add_91[0][0]               
#> ________________________________________________________________________________
#> conv3d_105 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_41[0][0] 
#> ________________________________________________________________________________
#> add_92 (Add)              (None, 16, 16, 16 0        conv3d_106[0][0]           
#>                                                      conv3d_105[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_709 ( (None, 16, 16, 16 512      add_92[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_42 (Thr (None, 16, 16, 16 0        batch_normalization_709[0][
#> ________________________________________________________________________________
#> conv3d_107 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_42[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_710 ( (None, 16, 16, 16 128      conv3d_107[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_43 (Thr (None, 16, 16, 16 0        batch_normalization_710[0][
#> ________________________________________________________________________________
#> conv3d_109 (Conv3D)       (None, 16, 16, 16 16512    add_92[0][0]               
#> ________________________________________________________________________________
#> conv3d_108 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_43[0][0] 
#> ________________________________________________________________________________
#> add_93 (Add)              (None, 16, 16, 16 0        conv3d_109[0][0]           
#>                                                      conv3d_108[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_711 ( (None, 16, 16, 16 512      add_93[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_44 (Thr (None, 16, 16, 16 0        batch_normalization_711[0][
#> ________________________________________________________________________________
#> conv3d_110 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_44[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_712 ( (None, 16, 16, 16 128      conv3d_110[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_45 (Thr (None, 16, 16, 16 0        batch_normalization_712[0][
#> ________________________________________________________________________________
#> conv3d_112 (Conv3D)       (None, 16, 16, 16 16512    add_93[0][0]               
#> ________________________________________________________________________________
#> conv3d_111 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_45[0][0] 
#> ________________________________________________________________________________
#> add_94 (Add)              (None, 16, 16, 16 0        conv3d_112[0][0]           
#>                                                      conv3d_111[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_713 ( (None, 16, 16, 16 512      add_94[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_46 (Thr (None, 16, 16, 16 0        batch_normalization_713[0][
#> ________________________________________________________________________________
#> conv3d_113 (Conv3D)       (None, 8, 8, 8, 6 8256     thresholded_re_lu_46[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_714 ( (None, 8, 8, 8, 6 256      conv3d_113[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_47 (Thr (None, 8, 8, 8, 6 0        batch_normalization_714[0][
#> ________________________________________________________________________________
#> conv3d_114 (Conv3D)       (None, 8, 8, 8, 6 4160     thresholded_re_lu_47[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_715 ( (None, 8, 8, 8, 6 256      conv3d_114[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_48 (Thr (None, 8, 8, 8, 6 0        batch_normalization_715[0][
#> ________________________________________________________________________________
#> conv3d_116 (Conv3D)       (None, 8, 8, 8, 2 33024    add_94[0][0]               
#> ________________________________________________________________________________
#> conv3d_115 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_48[0][0] 
#> ________________________________________________________________________________
#> add_95 (Add)              (None, 8, 8, 8, 2 0        conv3d_116[0][0]           
#>                                                      conv3d_115[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_716 ( (None, 8, 8, 8, 2 1024     add_95[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_49 (Thr (None, 8, 8, 8, 2 0        batch_normalization_716[0][
#> ________________________________________________________________________________
#> conv3d_117 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_49[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_717 ( (None, 8, 8, 8, 6 256      conv3d_117[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_50 (Thr (None, 8, 8, 8, 6 0        batch_normalization_717[0][
#> ________________________________________________________________________________
#> conv3d_119 (Conv3D)       (None, 8, 8, 8, 2 65792    add_95[0][0]               
#> ________________________________________________________________________________
#> conv3d_118 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_50[0][0] 
#> ________________________________________________________________________________
#> add_96 (Add)              (None, 8, 8, 8, 2 0        conv3d_119[0][0]           
#>                                                      conv3d_118[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_718 ( (None, 8, 8, 8, 2 1024     add_96[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_51 (Thr (None, 8, 8, 8, 2 0        batch_normalization_718[0][
#> ________________________________________________________________________________
#> conv3d_120 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_51[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_719 ( (None, 8, 8, 8, 6 256      conv3d_120[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_52 (Thr (None, 8, 8, 8, 6 0        batch_normalization_719[0][
#> ________________________________________________________________________________
#> conv3d_122 (Conv3D)       (None, 8, 8, 8, 2 65792    add_96[0][0]               
#> ________________________________________________________________________________
#> conv3d_121 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_52[0][0] 
#> ________________________________________________________________________________
#> add_97 (Add)              (None, 8, 8, 8, 2 0        conv3d_122[0][0]           
#>                                                      conv3d_121[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_720 ( (None, 8, 8, 8, 2 1024     add_97[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_53 (Thr (None, 8, 8, 8, 2 0        batch_normalization_720[0][
#> ________________________________________________________________________________
#> conv3d_123 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_53[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_721 ( (None, 8, 8, 8, 6 256      conv3d_123[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_54 (Thr (None, 8, 8, 8, 6 0        batch_normalization_721[0][
#> ________________________________________________________________________________
#> conv3d_125 (Conv3D)       (None, 8, 8, 8, 2 65792    add_97[0][0]               
#> ________________________________________________________________________________
#> conv3d_124 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_54[0][0] 
#> ________________________________________________________________________________
#> add_98 (Add)              (None, 8, 8, 8, 2 0        conv3d_125[0][0]           
#>                                                      conv3d_124[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_722 ( (None, 8, 8, 8, 2 1024     add_98[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_55 (Thr (None, 8, 8, 8, 2 0        batch_normalization_722[0][
#> ________________________________________________________________________________
#> conv3d_126 (Conv3D)       (None, 4, 4, 4, 1 32896    thresholded_re_lu_55[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_723 ( (None, 4, 4, 4, 1 512      conv3d_126[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_56 (Thr (None, 4, 4, 4, 1 0        batch_normalization_723[0][
#> ________________________________________________________________________________
#> conv3d_127 (Conv3D)       (None, 4, 4, 4, 1 16512    thresholded_re_lu_56[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_724 ( (None, 4, 4, 4, 1 512      conv3d_127[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_57 (Thr (None, 4, 4, 4, 1 0        batch_normalization_724[0][
#> ________________________________________________________________________________
#> conv3d_129 (Conv3D)       (None, 4, 4, 4, 5 131584   add_98[0][0]               
#> ________________________________________________________________________________
#> conv3d_128 (Conv3D)       (None, 4, 4, 4, 5 66048    thresholded_re_lu_57[0][0] 
#> ________________________________________________________________________________
#> add_99 (Add)              (None, 4, 4, 4, 5 0        conv3d_129[0][0]           
#>                                                      conv3d_128[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_725 ( (None, 4, 4, 4, 5 2048     add_99[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_58 (Thr (None, 4, 4, 4, 5 0        batch_normalization_725[0][
#> ________________________________________________________________________________
#> conv3d_130 (Conv3D)       (None, 4, 4, 4, 1 65664    thresholded_re_lu_58[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_726 ( (None, 4, 4, 4, 1 512      conv3d_130[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_59 (Thr (None, 4, 4, 4, 1 0        batch_normalization_726[0][
#> ________________________________________________________________________________
#> conv3d_transpose_3 (Conv3 (None, 4, 4, 4, 1 131200   thresholded_re_lu_59[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_4 (Conv3 (None, 4, 4, 4, 5 262656   add_99[0][0]               
#> ________________________________________________________________________________
#> up_sampling3d_5 (UpSampli (None, 8, 8, 8, 1 0        conv3d_transpose_3[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_6 (UpSampli (None, 8, 8, 8, 5 0        conv3d_transpose_4[0][0]   
#> ________________________________________________________________________________
#> conv3d_131 (Conv3D)       (None, 8, 8, 8, 5 66048    up_sampling3d_5[0][0]      
#> ________________________________________________________________________________
#> add_100 (Add)             (None, 8, 8, 8, 5 0        up_sampling3d_6[0][0]      
#>                                                      conv3d_131[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_727 ( (None, 8, 8, 8, 5 2048     add_100[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_60 (Thr (None, 8, 8, 8, 5 0        batch_normalization_727[0][
#> ________________________________________________________________________________
#> conv3d_132 (Conv3D)       (None, 8, 8, 8, 6 32832    thresholded_re_lu_60[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_728 ( (None, 8, 8, 8, 6 256      conv3d_132[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_61 (Thr (None, 8, 8, 8, 6 0        batch_normalization_728[0][
#> ________________________________________________________________________________
#> conv3d_134 (Conv3D)       (None, 8, 8, 8, 2 131328   add_100[0][0]              
#> ________________________________________________________________________________
#> conv3d_133 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_61[0][0] 
#> ________________________________________________________________________________
#> add_101 (Add)             (None, 8, 8, 8, 2 0        conv3d_134[0][0]           
#>                                                      conv3d_133[0][0]           
#> ________________________________________________________________________________
#> conv3d_135 (Conv3D)       (None, 8, 8, 8, 2 65792    add_101[0][0]              
#> ________________________________________________________________________________
#> add_102 (Add)             (None, 8, 8, 8, 2 0        add_98[0][0]               
#>                                                      conv3d_135[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_729 ( (None, 8, 8, 8, 2 1024     add_102[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_62 (Thr (None, 8, 8, 8, 2 0        batch_normalization_729[0][
#> ________________________________________________________________________________
#> conv3d_136 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_62[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_730 ( (None, 8, 8, 8, 6 256      conv3d_136[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_63 (Thr (None, 8, 8, 8, 6 0        batch_normalization_730[0][
#> ________________________________________________________________________________
#> conv3d_138 (Conv3D)       (None, 8, 8, 8, 2 65792    add_102[0][0]              
#> ________________________________________________________________________________
#> conv3d_137 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_63[0][0] 
#> ________________________________________________________________________________
#> add_103 (Add)             (None, 8, 8, 8, 2 0        conv3d_138[0][0]           
#>                                                      conv3d_137[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_731 ( (None, 8, 8, 8, 2 1024     add_103[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_64 (Thr (None, 8, 8, 8, 2 0        batch_normalization_731[0][
#> ________________________________________________________________________________
#> conv3d_139 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_64[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_732 ( (None, 8, 8, 8, 6 256      conv3d_139[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_65 (Thr (None, 8, 8, 8, 6 0        batch_normalization_732[0][
#> ________________________________________________________________________________
#> conv3d_141 (Conv3D)       (None, 8, 8, 8, 2 65792    add_103[0][0]              
#> ________________________________________________________________________________
#> conv3d_140 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_65[0][0] 
#> ________________________________________________________________________________
#> add_104 (Add)             (None, 8, 8, 8, 2 0        conv3d_141[0][0]           
#>                                                      conv3d_140[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_733 ( (None, 8, 8, 8, 2 1024     add_104[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_66 (Thr (None, 8, 8, 8, 2 0        batch_normalization_733[0][
#> ________________________________________________________________________________
#> conv3d_142 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_66[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_734 ( (None, 8, 8, 8, 6 256      conv3d_142[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_67 (Thr (None, 8, 8, 8, 6 0        batch_normalization_734[0][
#> ________________________________________________________________________________
#> conv3d_transpose_5 (Conv3 (None, 8, 8, 8, 6 32832    thresholded_re_lu_67[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_6 (Conv3 (None, 8, 8, 8, 2 65792    add_104[0][0]              
#> ________________________________________________________________________________
#> up_sampling3d_7 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_5[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_8 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_6[0][0]   
#> ________________________________________________________________________________
#> conv3d_143 (Conv3D)       (None, 16, 16, 16 16640    up_sampling3d_7[0][0]      
#> ________________________________________________________________________________
#> add_105 (Add)             (None, 16, 16, 16 0        up_sampling3d_8[0][0]      
#>                                                      conv3d_143[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_735 ( (None, 16, 16, 16 1024     add_105[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_68 (Thr (None, 16, 16, 16 0        batch_normalization_735[0][
#> ________________________________________________________________________________
#> conv3d_144 (Conv3D)       (None, 16, 16, 16 8224     thresholded_re_lu_68[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_736 ( (None, 16, 16, 16 128      conv3d_144[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_69 (Thr (None, 16, 16, 16 0        batch_normalization_736[0][
#> ________________________________________________________________________________
#> conv3d_146 (Conv3D)       (None, 16, 16, 16 32896    add_105[0][0]              
#> ________________________________________________________________________________
#> conv3d_145 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_69[0][0] 
#> ________________________________________________________________________________
#> add_106 (Add)             (None, 16, 16, 16 0        conv3d_146[0][0]           
#>                                                      conv3d_145[0][0]           
#> ________________________________________________________________________________
#> conv3d_147 (Conv3D)       (None, 16, 16, 16 16512    add_106[0][0]              
#> ________________________________________________________________________________
#> add_107 (Add)             (None, 16, 16, 16 0        add_94[0][0]               
#>                                                      conv3d_147[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_737 ( (None, 16, 16, 16 512      add_107[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_70 (Thr (None, 16, 16, 16 0        batch_normalization_737[0][
#> ________________________________________________________________________________
#> conv3d_148 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_70[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_738 ( (None, 16, 16, 16 128      conv3d_148[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_71 (Thr (None, 16, 16, 16 0        batch_normalization_738[0][
#> ________________________________________________________________________________
#> conv3d_150 (Conv3D)       (None, 16, 16, 16 16512    add_107[0][0]              
#> ________________________________________________________________________________
#> conv3d_149 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_71[0][0] 
#> ________________________________________________________________________________
#> add_108 (Add)             (None, 16, 16, 16 0        conv3d_150[0][0]           
#>                                                      conv3d_149[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_739 ( (None, 16, 16, 16 512      add_108[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_72 (Thr (None, 16, 16, 16 0        batch_normalization_739[0][
#> ________________________________________________________________________________
#> conv3d_151 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_72[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_740 ( (None, 16, 16, 16 128      conv3d_151[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_73 (Thr (None, 16, 16, 16 0        batch_normalization_740[0][
#> ________________________________________________________________________________
#> conv3d_transpose_7 (Conv3 (None, 16, 16, 16 8224     thresholded_re_lu_73[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_8 (Conv3 (None, 16, 16, 16 16512    add_108[0][0]              
#> ________________________________________________________________________________
#> up_sampling3d_9 (UpSampli (None, 32, 32, 32 0        conv3d_transpose_7[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_10 (UpSampl (None, 32, 32, 32 0        conv3d_transpose_8[0][0]   
#> ________________________________________________________________________________
#> conv3d_152 (Conv3D)       (None, 32, 32, 32 4224     up_sampling3d_9[0][0]      
#> ________________________________________________________________________________
#> add_109 (Add)             (None, 32, 32, 32 0        up_sampling3d_10[0][0]     
#>                                                      conv3d_152[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_741 ( (None, 32, 32, 32 512      add_109[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_74 (Thr (None, 32, 32, 32 0        batch_normalization_741[0][
#> ________________________________________________________________________________
#> conv3d_153 (Conv3D)       (None, 32, 32, 32 110624   thresholded_re_lu_74[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_10 (Conv (None, 32, 32, 32 4128     add_109[0][0]              
#> ________________________________________________________________________________
#> conv3d_transpose_9 (Conv3 (None, 32, 32, 32 8224     conv3d_153[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_12 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_10[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_11 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_9[0][0]   
#> ________________________________________________________________________________
#> add_110 (Add)             (None, 64, 64, 64 0        up_sampling3d_12[0][0]     
#>                                                      up_sampling3d_11[0][0]     
#> ________________________________________________________________________________
#> conv3d_154 (Conv3D)       (None, 64, 64, 64 27680    add_110[0][0]              
#> ________________________________________________________________________________
#> add_111 (Add)             (None, 64, 64, 64 0        conv3d_100[0][0]           
#>                                                      conv3d_154[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_742 ( (None, 64, 64, 64 128      add_111[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_75 (Thr (None, 64, 64, 64 0        batch_normalization_742[0][
#> ________________________________________________________________________________
#> conv3d_155 (Conv3D)       (None, 64, 64, 64 33       thresholded_re_lu_75[0][0] 
#> ================================================================================
#> Total params: 2,088,289
#> Trainable params: 2,078,049
#> Non-trainable params: 10,240
#> ________________________________________________________________________________
#> 
#>