2-D implementation of the spatial transform network.

Creates a keras model of the spatial transform network:

createSpatialTransformNetwork2D(inputImageSize, resampledSize = c(10,
  10), numberOfClassificationLabels = 10)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfClassificationLabels
Number of classes.

samplingSize
a list of 2-D vectors specifying the kernel size at each convolution layer. Default values are the same as given in the original paper. The length of kernel size vectors must be 1 greater than the vector length of the number of filters.

Value

a keras model for image super resolution

Details

https://arxiv.org/abs/1506.02025

based on the following python Keras model:

https://github.com/oarriaga/STN.keras/blob/master/src/models/STN.py

Examples

library( ANTsRNet )

mnistData <- dataset_mnist()

#> Error in dataset_mnist(): could not find function "dataset_mnist"

numberOfLabels <- 10

# Extract a small subset for something that can run quickly

X_trainSmall <- mnistData$train$x[1:100,,]

#> Error in eval(expr, envir, enclos): object 'mnistData' not found

X_trainSmall <- array( data = X_trainSmall, dim = c( dim( X_trainSmall ), 1 ) )

#> Error in array(data = X_trainSmall, dim = c(dim(X_trainSmall), 1)): object 'X_trainSmall' not found

Y_trainSmall <- to_categorical( mnistData$train$y[1:100], numberOfLabels )

#> Error in to_categorical(mnistData$train$y[1:100], numberOfLabels): could not find function "to_categorical"

X_testSmall <- mnistData$test$x[1:10,,]

#> Error in eval(expr, envir, enclos): object 'mnistData' not found

X_testSmall <- array( data = X_testSmall, dim = c( dim( X_testSmall ), 1 ) )

#> Error in array(data = X_testSmall, dim = c(dim(X_testSmall), 1)): object 'X_testSmall' not found

Y_testSmall <- to_categorical( mnistData$test$y[1:10], numberOfLabels )

#> Error in to_categorical(mnistData$test$y[1:10], numberOfLabels): could not find function "to_categorical"

# We add a dimension of 1 to specify the channel size

inputImageSize <- c( dim( X_trainSmall )[2:3], 1 )

#> Error in eval(expr, envir, enclos): object 'X_trainSmall' not found

model <- createSpatialTransformNetworkModel2D( inputImageSize = inputImageSize,
  numberOfClassificationLabels = numberOfLabels )

#> Error in createSpatialTransformNetworkModel2D(inputImageSize = inputImageSize,     numberOfClassificationLabels = numberOfLabels): could not find function "createSpatialTransformNetworkModel2D"