2-D implementation of the Vgg deep learning architecture.

Creates a keras model of the Vgg deep learning architecture for image recognition based on the paper

createVggModel2D(inputImageSize, numberOfClassificationLabels = 1000,
  layers = c(1, 2, 3, 4, 4), lowestResolution = 64,
  convolutionKernelSize = c(3, 3), poolSize = c(2, 2), strides = c(2,
  2), denseUnits = 4096, dropoutRate = 0, style = 19,
  mode = "classification")

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfClassificationLabels
Number of segmentation labels.

layers
a vector determining the number of filters defined at for each layer.

lowestResolution
number of filters at the beginning.

convolutionKernelSize
2-d vector definining the kernel size during the encoding path

poolSize
2-d vector defining the region for each pooling layer.

strides
2-d vector describing the stride length in each direction.

denseUnits
integer for the number of units in the last layers.

dropoutRate
float between 0 and 1 to use between dense layers.

style
'16' or '19' for VGG16 or VGG19, respectively.

mode
'classification' or 'regression'. Default = 'classification'.

Value

a VGG keras model to be used with subsequent fitting

Details

K. Simonyan and A. Zisserman, Very Deep Convolutional Networks for Large-Scale Image Recognition

available here:

https://arxiv.org/abs/1409.1556

This particular implementation was influenced by the following python implementation:

https://gist.github.com/baraldilorenzo/8d096f48a1be4a2d660d

Examples

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )
library( keras )

mnistData <- dataset_mnist()
numberOfLabels <- 10

# Extract a small subset for something that can run quickly

X_trainSmall <- mnistData$train$x[1:10,,]
X_trainSmall <- array( data = X_trainSmall, dim = c( dim( X_trainSmall ), 1 ) )
Y_trainSmall <- to_categorical( mnistData$train$y[1:10], numberOfLabels )

X_testSmall <- mnistData$test$x[1:10,,]
X_testSmall <- array( data = X_testSmall, dim = c( dim( X_testSmall ), 1 ) )
Y_testSmall <- to_categorical( mnistData$test$y[1:10], numberOfLabels )

# We add a dimension of 1 to specify the channel size

inputImageSize <- c( dim( X_trainSmall )[2:3], 1 )

model <- createVggModel2D( inputImageSize = inputImageSize,
  numberOfClassificationLabels = numberOfLabels )

#> Error in py_call_impl(callable, dots$args, dots$keywords): ValueError: Negative dimension size caused by subtracting 2 from 1 for 'max_pooling2d_11/MaxPool' (op: 'MaxPool') with input shapes: [?,1,1,512].
#> 
#> Detailed traceback: 
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/engine/sequential.py", line 185, in add
#>     output_tensor = layer(self.outputs[0])
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/engine/base_layer.py", line 457, in __call__
#>     output = self.call(inputs, **kwargs)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/layers/pooling.py", line 157, in call
#>     data_format=self.data_format)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/layers/pooling.py", line 220, in _pooling_function
#>     pool_mode='max')
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/backend/tensorflow_backend.py", line 3880, in pool2d
#>     data_format=tf_data_format)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/nn_ops.py", line 2144, in max_pool
#>     name=name)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/gen_nn_ops.py", line 4587, in max_pool
#>     data_format=data_format, name=name)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/op_def_library.py", line 787, in _apply_op_helper
#>     op_def=op_def)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 3292, in create_op
#>     compute_device=compute_device)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 3332, in _create_op_helper
#>     set_shapes_for_outputs(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2496, in set_shapes_for_outputs
#>     return _set_shapes_for_outputs(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2469, in _set_shapes_for_outputs
#>     shapes = shape_func(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2399, in call_with_requiring
#>     return call_cpp_shape_fn(op, require_shape_fn=True)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/common_shapes.py", line 627, in call_cpp_shape_fn
#>     require_shape_fn)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/common_shapes.py", line 691, in _call_cpp_shape_fn_impl
#>     raise ValueError(err.message)

model %>% compile( loss = 'categorical_crossentropy',
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( 'categorical_crossentropy', 'accuracy' ) )

#> Error in eval(lhs, parent, parent): object 'model' not found

# Comment out the rest due to travis build constraints

# track <- model %>% fit( X_trainSmall, Y_trainSmall, verbose = 1,
#   epochs = 1, batch_size = 2, shuffle = TRUE, validation_split = 0.5 )

# Now test the model

# testingMetrics <- model %>% evaluate( X_testSmall, Y_testSmall )
# predictedData <- model %>% predict( X_testSmall, verbose = 1 )