Randomly transform image data.

Given an input image list (possibly multi-modal) and an optional corresponding segmentation image list, this function will perform data augmentation with the following augmentation possibilities: spatial transformations, added image noise, simulated bias field, and histogram warping.

dataAugmentation(
  inputImageList,
  segmentationImageList = NULL,
  numberOfSimulations = 10,
  referenceImage = NULL,
  transformType = "affineAndDeformation",
  noiseModel = "additivegaussian",
  noiseParameters = c(0, 0.05),
  sdSimulatedBiasField = 0.05,
  sdHistogramWarping = 0.05,
  outputNumpyFilePrefix = NULL,
  verbose = FALSE
)

Arguments

inputImageList
list of lists of input images to warp. The internal list sets contains one or more images (per subject) which are assumed to be mutually aligned. The outer list contains multiple subject lists which are randomly sampled to produce output image list.

segmentationImageList
list of segmentation images corresponding to the input image list (optional).

numberOfSimulations
number of output images. Default = 10.

referenceImage
defines the spatial domain for all output images. If the input images do not match the spatial domain of the reference image, we internally resample the target to the reference image. This could have unexpected consequences. Resampling to the reference domain is performed by testing using antsImagePhysicalSpaceConsistency then calling resampleImageToTarget upon failure.

transformType
one of the following options c( "translation", "rigid", "scaleShear", "affine"," deformation" , "affineAndDeformation" ).

noiseModel
one of the following options c( "additivegaussian", "saltandpepper", "shot", "speckle" )

noiseParameters
'additivegaussian': c( mean, standardDeviation ), 'saltandpepper': c( probability, saltValue, pepperValue) , 'shot': scale, 'speckle': standardDeviation. Note that the standard deviation, scale, and probability values are max values and are randomly selected in the range 0, noise_parameter. Also, the "mean", "saltValue" and pepperValue" are assumed to be in the intensity normalized range of 0, 1.

sdSimulatedBiasField
Characterize the standard deviation of the amplitude.

sdHistogramWarping
Determines the strength of the bias field.

outputNumpyFilePrefix
Filename of output numpy array containing all the simulated images and segmentations.

Value

list of lists of transformed images and/or outputs to a numpy array.

Examples

library( ANTsR )
image1 <- antsImageRead( getANTsRData( "r16" ) )
image2 <- antsImageRead( getANTsRData( "r64" ) )
segmentation1 <- thresholdImage( image1, "Otsu", 3 )
segmentation2 <- thresholdImage( image2, "Otsu", 3 )
data <- dataAugmentation(
  list( list( image1 ), list( image2 ) ),
  list( segmentation1, segmentation2 ) )

#>         Adding simulated bias field.
#>         Performing intensity histogram warping.

#> Error in dataAugmentation(list(list(image1), list(image2)), list(segmentation1,     segmentation2)): object 'break_points' not found

rm(segmentation1); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2550453 136.3    4601961 245.8         NA   4601961 245.8
#> Vcells 6914123  52.8   44606902 340.4     102400 127497723 972.8

rm(segmentation2); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2549694 136.2    4601961 245.8         NA   4601961 245.8
#> Vcells 6912894  52.8   35685522 272.3     102400 127497723 972.8

rm(image1); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2549685 136.2    4601961 245.8         NA   4601961 245.8
#> Vcells 6912915  52.8   28548418 217.9     102400 127497723 972.8

rm(image2); gc()

#>           used  (Mb) gc trigger  (Mb) limit (Mb)  max used  (Mb)
#> Ncells 2549675 136.2    4601961 245.8         NA   4601961 245.8
#> Vcells 6912920  52.8   28548418 217.9     102400 127497723 972.8