Hippocampal/Enthorhinal segmentation using "Deep Flash"

Perform hippocampal/entorhinal segmentation in T1 images using labels from Mike Yassa's lab

deepFlash(
  t1,
  t2 = NULL,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  useRankIntensity = FALSE,
  verbose = FALSE
)

Arguments

t1
raw or preprocessed 3-D T1-weighted brain image.

t2
optional raw or preprocessed 3-D T2-weighted brain image.

doPreprocessing
perform preprocessing. See description above.

antsxnetCacheDirectory
destination directory for storing the downloaded template and model weights. Since these can be resused, if is.null(antsxnetCacheDirectory), these data will be downloaded to the inst/extdata/ subfolder of the ANTsRNet package.

useRankIntensity
If false, use histogram matching with cropped template ROI. Otherwise, use a rank intensity transform on the cropped ROI.

verbose
print progress.

Value

list consisting of the segmentation image and probability images for each label.

Details

https://faculty.sites.uci.edu/myassa/

The labeling is as follows:

Label 0 :background

Label 5 :left aLEC

Label 6 :right aLEC

Label 7 :left pMEC

Label 8 :right pMEC

Label 9 :left perirhinal

Label 10:right perirhinal

Label 11:left parahippocampal

Label 12:right parahippocampal

Label 13:left DG/CA2/CA3/CA4

Label 14:right DG/CA2/CA3/CA4

Label 15:left CA1

Label 16:right CA1

Label 17:left subiculum

Label 18:right subiculum

Preprocessing on the training data consisted of:

n4 bias correction,

affine registration to deep flash template. The input T1 should undergo the same steps. If the input T1 is the raw T1, these steps can be performed by the internal preprocessing, i.e. set doPreprocessing = TRUE

Examples

if (FALSE) {
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- deepFlash( image )
}