Extract 2-D or 3-D image patch coordinates.

Extract 2-D or 3-D image patch coordinates.

extractImagePatchCoordinates(
  image,
  patchSize,
  maxNumberOfPatches = "all",
  strideLength = 1,
  maskImage = NULL,
  physicalCoordinates = TRUE,
  cornerCoordinates = TRUE,
  randomSeed,
  randomize = TRUE
)

Arguments

image
Input ANTs image with one or more components.

patchSize
Width, height, and depth (if 3-D) of patches.

maxNumberOfPatches
Maximum number of patches returned. If "all" is specified, then all patches in sequence (defined by the strideLength are extracted.

strideLength
Defines the sequential patch overlap for maxNumberOfPatches = "all". Can be a image-dimensional vector or a scalar.

maskImage
optional image specifying the sampling region for the patches when maximumNumberOfPatches does not equal "all". The way we constrain patch selection using a mask is by forcing each returned patch to have a masked voxel at its center.

physicalCoordinates
boolean to determine whether indices or spatial coordinates are returned.

cornerCoordinates
boolean to determine whether indices or spatial coordinates of the corner or the center are returned.

randomSeed
integer seed that allows reproducible patch extraction across runs.

randomize
boolean controlling whether we randomize indices when masking.

Value

a matrix of image patch spatial or index coordinates.

Examples

library( ANTsR )

image <- antsImageRead( getANTsRData( "r16" ) )
maskImage <- getMask( image, 1, 1000 )
patchCoordsP <- extractImagePatchCoordinates( image, c( 32, 32 ), 10, c( 32, 32 ),
  maskImage, randomSeed = 0, physicalCoordinates = TRUE )
patchCoordsI <- extractImagePatchCoordinates( image, c( 32, 32 ), 10, c( 32, 32 ),
  maskImage, randomSeed = 0, physicalCoordinates = FALSE )