Extract 2-D or 3-D image patches.

Extract 2-D or 3-D image patches.

extractImagePatches(image, patchSize, maxNumberOfPatches = "all",
  strideLength = 1, maskImage = NA, randomSeed)

Arguments

image
Input ANTs image

patchSize
Width, height, and depth (if 3-D) of patches.

maxNumberOfPatches
Maximum number of patches returned. If "all" is specified, then all patches in sequence (defined by the

strideLength
Defines the sequential patch overlap for maxNumberOfPatches = all. Can be a image-dimensional vector or a scalar.

maskImage
optional image specifying the sampling region for the patches when maximumNumberOfPatches does not equal "all".

randomSeed
integer seed that allows reproducible patch extraction across runs.

Value

a randomly selected list of patches.

Examples

library( ANTsR )

image <- antsImageRead( getANTsRData( "r16" ) )
maskImage <- getMask( image, 1, 1000 )
patchSet1 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 0 )
patchSet2 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 1 )
patchSet3 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), maskImage, randomSeed = 0 )