hippMapp3rSegmentation

Perform HippMapp3r (hippocampal) segmentation described in

hippMapp3rSegmentation(
  t1,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)

Arguments

doPreprocessing
perform preprocessing. See description above.

antsxnetCacheDirectory
destination directory for storing the downloaded template and model weights. Since these can be resused, if is.null(antsxnetCacheDirectory), these data will be downloaded to the subdirectory ~/.keras/ANTsXNet/.

verbose
print progress.

image
input 3-D T1-weighted brain image.

Value

labeled hippocampal mask (ANTsR image)

Details

https://www.ncbi.nlm.nih.gov/pubmed/31609046

with models and architecture ported from

https://github.com/mgoubran/HippMapp3r

Additional documentation and attribution resources found at

https://hippmapp3r.readthedocs.io/en/latest/

Preprocessing consists of:

n4 bias correction and

brain extraction. The input T1 should undergo the same steps. If the input T1 is the raw T1, these steps can be performed by the internal preprocessing, i.e. set doPreprocessing = TRUE

Author

Tustison NJ

Examples

if (FALSE) {
library( ANTsRNet )
library( ANTsRNet )
library( keras )

url <- "https://github.com/mgoubran/HippMapp3r/blob/master/data/test_case/mprage.nii.gz?raw=true"
imageFile <- "head.nii.gz"
download.file( url, imageFile )
image <- antsImageRead( imageFile )
imageN4 <- n4BiasFieldCorrection( image, verbose = TRUE )
segmentation <- hippMapp3rSegmentation( imageN4, verbose = TRUE )
}