Instance normalization layer

Creates an instance normalization layer

layer_instance_normalization(object, axis = -1L, epsilon = 0.001,
  center = TRUE, scale = TRUE, betaInitializer = "zeros",
  gammaInitializer = "ones", betaRegularizer = NULL,
  gammaRegularizer = NULL, betaConstraint = NULL,
  gammaConstraint = NULL, trainable = TRUE)

Arguments

axis
Integer specifying which axis should be normalized, typically the feature axis. For example, after a Conv2D layer with channels_first, set axis = 1. Setting axis=-1L will normalize all values in each instance of the batch. Axis 0 is the batch dimension for tensorflow backend so we throw an error if axis = 0.

epsilon
Small float added to the variance to avoid dividing by 0.

center
If TRUE, add beta offset to normalized tensor.

scale
If TRUE, multiply by gamma.

betaInitializer
Intializer for the beta weight.

gammaInitializer
Intializer for the gamma weight.

betaRegularizer
Regularizer for the beta weight.

gammaRegularizer
Regularizer for the gamma weight.

betaConstraint
Optional constraint for the beta weight.

gammaConstraint
Optional constraint for the gamma weight.

Value

a keras layer tensor