Loss function for the SSD deep learning architecture.

Creates an R6 class object for use with the SSD deep learning architecture based on the paper

Value

an SSD loss function

Details

W. Liu, D. Anguelov, D. Erhan, C. Szegedy, S. Reed, C-Y. Fu, A. Berg. SSD: Single Shot MultiBox Detector.

available here:

    \url{https://arxiv.org/abs/1512.02325}

Usage

ssdLoss &lt;- LossSSD$new( dimension = 2L, backgroundRatio = 3L,
  minNumberOfBackgroundBoxes = 0L,  alpha = 1.0,
  numberOfClassificationLabels )

ssdLoss$smooth_l1_loss( y_true, y_pred )
ssdLoss$log_loss( y_true, y_pred )
ssdLoss$compute_loss( y_true, y_pred )

Arguments

ssdLoss

A process object.

dimension

image dimensionality.

backgroundRatio

The maximum ratio of background to foreround for weighting in the loss function. Is rounded to the nearest integer. Default is 3.

minNumberOfBackgroundBoxes

The minimum number of background boxes to use in loss computation per batch. Should reflect a value in proportion to the batch size. Default is 0.

alpha

Weighting factor for the localization loss in total loss computation.

numberOfClassificationLabels

number of classes including background.

Details

$smooth_l1_loss smooth loss

$log_loss log loss

$compute_loss computes total loss.