2-D implementation of the GoogLeNet deep learning architecture.

Creates a keras model of the GoogLeNet deep learning architecture for image recognition based on the paper

createGoogLeNetModel2D(inputImageSize,
  numberOfClassificationLabels = 1000, mode = "classification")

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfClassificationLabels
Number of segmentation labels.

mode
'classification' or 'regression'. Default = 'classification'.

Value

a GoogLeNet keras model

Details

C. Szegedy, W. Liu, Y. Jia, P. Sermanet, S. Reed, D. Anguelov, D. Erhan, V. Vanhoucke, A. Rabinovich, Going Deeper with Convolutions C. Szegedy, V. Vanhoucke, S. Ioffe, J. Shlens, and Z. Wojna. Rethinking the Inception Architecture for Computer Vision

available here:

https://arxiv.org/abs/1409.4842 https://arxiv.org/abs/1512.00567

This particular implementation was influenced by the following python implementation:

https://github.com/fchollet/deep-learning-models/blob/master/inception_v3.py

Examples

library( ANTsRNet )
library( keras )

mnistData <- dataset_mnist()
numberOfLabels <- 10

# Extract a small subset for something that can run quickly

X_trainSmall <- mnistData$train$x[1:10,,]
X_trainSmall <- array( data = X_trainSmall, dim = c( dim( X_trainSmall ), 1 ) )
Y_trainSmall <- to_categorical( mnistData$train$y[1:10], numberOfLabels )

X_testSmall <- mnistData$test$x[1:10,,]
X_testSmall <- array( data = X_testSmall, dim = c( dim( X_testSmall ), 1 ) )
Y_testSmall <- to_categorical( mnistData$test$y[1:10], numberOfLabels )

# We add a dimension of 1 to specify the channel size

inputImageSize <- c( dim( X_trainSmall )[2:3], 1 )

model <- createGoogLeNetModel2D( inputImageSize = inputImageSize,
  numberOfClassificationLabels = numberOfLabels )

#> Error in py_call_impl(callable, dots$args, dots$keywords): ValueError: Negative dimension size caused by subtracting 3 from 2 for 'conv2d_32/convolution' (op: 'Conv2D') with input shapes: [?,2,2,256], [3,3,256,384].
#> 
#> Detailed traceback: 
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/engine/base_layer.py", line 457, in __call__
#>     output = self.call(inputs, **kwargs)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/layers/convolutional.py", line 168, in call
#>     dilation_rate=self.dilation_rate)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/keras/backend/tensorflow_backend.py", line 3565, in conv2d
#>     data_format=tf_data_format)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/nn_ops.py", line 782, in convolution
#>     return op(input, filter)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/nn_ops.py", line 870, in __call__
#>     return self.conv_op(inp, filter)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/nn_ops.py", line 522, in __call__
#>     return self.call(inp, filter)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/nn_ops.py", line 206, in __call__
#>     name=self.name)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/ops/gen_nn_ops.py", line 953, in conv2d
#>     data_format=data_format, dilations=dilations, name=name)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/op_def_library.py", line 787, in _apply_op_helper
#>     op_def=op_def)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 3292, in create_op
#>     compute_device=compute_device)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 3332, in _create_op_helper
#>     set_shapes_for_outputs(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2496, in set_shapes_for_outputs
#>     return _set_shapes_for_outputs(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2469, in _set_shapes_for_outputs
#>     shapes = shape_func(op)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 2399, in call_with_requiring
#>     return call_cpp_shape_fn(op, require_shape_fn=True)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/common_shapes.py", line 627, in call_cpp_shape_fn
#>     require_shape_fn)
#>   File "/Users/ntustison/anaconda3/lib/python3.6/site-packages/tensorflow/python/framework/common_shapes.py", line 691, in _call_cpp_shape_fn_impl
#>     raise ValueError(err.message)

model %>% compile( loss = 'categorical_crossentropy',
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( 'categorical_crossentropy', 'accuracy' ) )

#> Error in eval(lhs, parent, parent): object 'model' not found

# Comment out the rest due to travis build constraints

# track <- model %>% fit( X_trainSmall, Y_trainSmall, verbose = 1,
#   epochs = 1, batch_size = 2, shuffle = TRUE, validation_split = 0.5 )

# Now test the model

# testingMetrics <- model %>% evaluate( X_testSmall, Y_testSmall )
# predictedData <- model %>% predict( X_testSmall, verbose = 1 )