2-D implementation of the ResNet image super resolution architecture.

Creates a keras model of the expanded image super resolution deep learning framework based on the following python implementation:

createResNetSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSize = c(3, 3),
  numberOfFilters = 64,
  numberOfResidualBlocks = 5,
  numberOfResNetBlocks = 1
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSize
a vector specifying the kernel size for convolution.

numberOfFilters
the number of filters for each encoding layer.

numberOfResidualBlocks
the number of residual blocks.

numberOfResNetBlocks
the number of resNet blocks. Each block will double the upsampling amount.

Value

a keras model for ResNet image super resolution

Details

    \url{https://github.com/titu1994/Image-Super-Resolution}

Examples

createResNetSuperResolutionModel2D(c(256, 256, 3))

#> Model
#> Model: "model_32"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_29 (InputLayer)     [(None, 256, 256, 0                                   
#> ________________________________________________________________________________
#> conv2d_639 (Conv2D)       (None, 256, 256,  1792     input_29[0][0]             
#> ________________________________________________________________________________
#> conv2d_640 (Conv2D)       (None, 256, 256,  36928    conv2d_639[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_468 ( (None, 256, 256,  256      conv2d_640[0][0]           
#> ________________________________________________________________________________
#> activation_349 (Activatio (None, 256, 256,  0        batch_normalization_468[0][
#> ________________________________________________________________________________
#> conv2d_641 (Conv2D)       (None, 256, 256,  36928    activation_349[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_469 ( (None, 256, 256,  256      conv2d_641[0][0]           
#> ________________________________________________________________________________
#> add_152 (Add)             (None, 256, 256,  0        conv2d_639[0][0]           
#>                                                      batch_normalization_469[0][
#> ________________________________________________________________________________
#> conv2d_642 (Conv2D)       (None, 256, 256,  36928    add_152[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_470 ( (None, 256, 256,  256      conv2d_642[0][0]           
#> ________________________________________________________________________________
#> activation_350 (Activatio (None, 256, 256,  0        batch_normalization_470[0][
#> ________________________________________________________________________________
#> conv2d_643 (Conv2D)       (None, 256, 256,  36928    activation_350[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_471 ( (None, 256, 256,  256      conv2d_643[0][0]           
#> ________________________________________________________________________________
#> add_153 (Add)             (None, 256, 256,  0        add_152[0][0]              
#>                                                      batch_normalization_471[0][
#> ________________________________________________________________________________
#> conv2d_644 (Conv2D)       (None, 256, 256,  36928    add_153[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_472 ( (None, 256, 256,  256      conv2d_644[0][0]           
#> ________________________________________________________________________________
#> activation_351 (Activatio (None, 256, 256,  0        batch_normalization_472[0][
#> ________________________________________________________________________________
#> conv2d_645 (Conv2D)       (None, 256, 256,  36928    activation_351[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_473 ( (None, 256, 256,  256      conv2d_645[0][0]           
#> ________________________________________________________________________________
#> add_154 (Add)             (None, 256, 256,  0        add_153[0][0]              
#>                                                      batch_normalization_473[0][
#> ________________________________________________________________________________
#> conv2d_646 (Conv2D)       (None, 256, 256,  36928    add_154[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_474 ( (None, 256, 256,  256      conv2d_646[0][0]           
#> ________________________________________________________________________________
#> activation_352 (Activatio (None, 256, 256,  0        batch_normalization_474[0][
#> ________________________________________________________________________________
#> conv2d_647 (Conv2D)       (None, 256, 256,  36928    activation_352[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_475 ( (None, 256, 256,  256      conv2d_647[0][0]           
#> ________________________________________________________________________________
#> add_155 (Add)             (None, 256, 256,  0        add_154[0][0]              
#>                                                      batch_normalization_475[0][
#> ________________________________________________________________________________
#> conv2d_648 (Conv2D)       (None, 256, 256,  36928    add_155[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_476 ( (None, 256, 256,  256      conv2d_648[0][0]           
#> ________________________________________________________________________________
#> activation_353 (Activatio (None, 256, 256,  0        batch_normalization_476[0][
#> ________________________________________________________________________________
#> conv2d_649 (Conv2D)       (None, 256, 256,  36928    activation_353[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_477 ( (None, 256, 256,  256      conv2d_649[0][0]           
#> ________________________________________________________________________________
#> add_156 (Add)             (None, 256, 256,  0        add_155[0][0]              
#>                                                      batch_normalization_477[0][
#> ________________________________________________________________________________
#> conv2d_650 (Conv2D)       (None, 256, 256,  36928    add_156[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_478 ( (None, 256, 256,  256      conv2d_650[0][0]           
#> ________________________________________________________________________________
#> activation_354 (Activatio (None, 256, 256,  0        batch_normalization_478[0][
#> ________________________________________________________________________________
#> conv2d_651 (Conv2D)       (None, 256, 256,  36928    activation_354[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_479 ( (None, 256, 256,  256      conv2d_651[0][0]           
#> ________________________________________________________________________________
#> add_157 (Add)             (None, 256, 256,  0        add_156[0][0]              
#>                                                      batch_normalization_479[0][
#> ________________________________________________________________________________
#> add_158 (Add)             (None, 256, 256,  0        add_157[0][0]              
#>                                                      conv2d_639[0][0]           
#> ________________________________________________________________________________
#> up_sampling2d_7 (UpSampli (None, 512, 512,  0        add_158[0][0]              
#> ________________________________________________________________________________
#> conv2d_652 (Conv2D)       (None, 512, 512,  36928    up_sampling2d_7[0][0]      
#> ________________________________________________________________________________
#> conv2d_653 (Conv2D)       (None, 512, 512,  1731     conv2d_652[0][0]           
#> ================================================================================
#> Total params: 486,659
#> Trainable params: 485,123
#> Non-trainable params: 1,536
#> ________________________________________________________________________________
#> 
#>