3-D implementation of the ResNet image super resolution architecture.

Creates a keras model of the expanded image super resolution deep learning framework based on the following python implementation:

createResNetSuperResolutionModel3D(
  inputImageSize,
  convolutionKernelSize = c(3, 3, 3),
  numberOfFilters = 64,
  numberOfResidualBlocks = 5,
  numberOfResNetBlocks = 1
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

convolutionKernelSize
a vector specifying the kernel size for convolution.

numberOfFilters
the number of filters for each encoding layer.

numberOfResidualBlocks
the number of residual blocks.

numberOfResNetBlocks
the number of resNet blocks. Each block will double the upsampling amount.

Value

a keras model for ResNet image super resolution

Details

    \url{https://github.com/titu1994/Image-Super-Resolution}

Author

Tustison NJ

Examples

inputImageSize = c(256, 256, 30, 1)
createResNetSuperResolutionModel3D(inputImageSize)

#> Model
#> Model: "model_34"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_31 (InputLayer)     [(None, 256, 256, 0                                   
#> ________________________________________________________________________________
#> conv3d_129 (Conv3D)       (None, 256, 256,  1792     input_31[0][0]             
#> ________________________________________________________________________________
#> conv3d_130 (Conv3D)       (None, 256, 256,  110656   conv3d_129[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_642 ( (None, 256, 256,  256      conv3d_130[0][0]           
#> ________________________________________________________________________________
#> activation_516 (Activatio (None, 256, 256,  0        batch_normalization_642[0][
#> ________________________________________________________________________________
#> conv3d_131 (Conv3D)       (None, 256, 256,  110656   activation_516[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_643 ( (None, 256, 256,  256      conv3d_131[0][0]           
#> ________________________________________________________________________________
#> add_162 (Add)             (None, 256, 256,  0        conv3d_129[0][0]           
#>                                                      batch_normalization_643[0][
#> ________________________________________________________________________________
#> conv3d_132 (Conv3D)       (None, 256, 256,  110656   add_162[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_644 ( (None, 256, 256,  256      conv3d_132[0][0]           
#> ________________________________________________________________________________
#> activation_517 (Activatio (None, 256, 256,  0        batch_normalization_644[0][
#> ________________________________________________________________________________
#> conv3d_133 (Conv3D)       (None, 256, 256,  110656   activation_517[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_645 ( (None, 256, 256,  256      conv3d_133[0][0]           
#> ________________________________________________________________________________
#> add_163 (Add)             (None, 256, 256,  0        add_162[0][0]              
#>                                                      batch_normalization_645[0][
#> ________________________________________________________________________________
#> conv3d_134 (Conv3D)       (None, 256, 256,  110656   add_163[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_646 ( (None, 256, 256,  256      conv3d_134[0][0]           
#> ________________________________________________________________________________
#> activation_518 (Activatio (None, 256, 256,  0        batch_normalization_646[0][
#> ________________________________________________________________________________
#> conv3d_135 (Conv3D)       (None, 256, 256,  110656   activation_518[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_647 ( (None, 256, 256,  256      conv3d_135[0][0]           
#> ________________________________________________________________________________
#> add_164 (Add)             (None, 256, 256,  0        add_163[0][0]              
#>                                                      batch_normalization_647[0][
#> ________________________________________________________________________________
#> conv3d_136 (Conv3D)       (None, 256, 256,  110656   add_164[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_648 ( (None, 256, 256,  256      conv3d_136[0][0]           
#> ________________________________________________________________________________
#> activation_519 (Activatio (None, 256, 256,  0        batch_normalization_648[0][
#> ________________________________________________________________________________
#> conv3d_137 (Conv3D)       (None, 256, 256,  110656   activation_519[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_649 ( (None, 256, 256,  256      conv3d_137[0][0]           
#> ________________________________________________________________________________
#> add_165 (Add)             (None, 256, 256,  0        add_164[0][0]              
#>                                                      batch_normalization_649[0][
#> ________________________________________________________________________________
#> conv3d_138 (Conv3D)       (None, 256, 256,  110656   add_165[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_650 ( (None, 256, 256,  256      conv3d_138[0][0]           
#> ________________________________________________________________________________
#> activation_520 (Activatio (None, 256, 256,  0        batch_normalization_650[0][
#> ________________________________________________________________________________
#> conv3d_139 (Conv3D)       (None, 256, 256,  110656   activation_520[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_651 ( (None, 256, 256,  256      conv3d_139[0][0]           
#> ________________________________________________________________________________
#> add_166 (Add)             (None, 256, 256,  0        add_165[0][0]              
#>                                                      batch_normalization_651[0][
#> ________________________________________________________________________________
#> conv3d_140 (Conv3D)       (None, 256, 256,  110656   add_166[0][0]              
#> ________________________________________________________________________________
#> batch_normalization_652 ( (None, 256, 256,  256      conv3d_140[0][0]           
#> ________________________________________________________________________________
#> activation_521 (Activatio (None, 256, 256,  0        batch_normalization_652[0][
#> ________________________________________________________________________________
#> conv3d_141 (Conv3D)       (None, 256, 256,  110656   activation_521[0][0]       
#> ________________________________________________________________________________
#> batch_normalization_653 ( (None, 256, 256,  256      conv3d_141[0][0]           
#> ________________________________________________________________________________
#> add_167 (Add)             (None, 256, 256,  0        add_166[0][0]              
#>                                                      batch_normalization_653[0][
#> ________________________________________________________________________________
#> add_168 (Add)             (None, 256, 256,  0        add_167[0][0]              
#>                                                      conv3d_129[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_3 (UpSampli (None, 512, 512,  0        add_168[0][0]              
#> ________________________________________________________________________________
#> conv3d_142 (Conv3D)       (None, 512, 512,  110656   up_sampling3d_3[0][0]      
#> ________________________________________________________________________________
#> conv3d_143 (Conv3D)       (None, 512, 512,  1729     conv3d_142[0][0]           
#> ================================================================================
#> Total params: 1,445,121
#> Trainable params: 1,443,585
#> Non-trainable params: 1,536
#> ________________________________________________________________________________
#> 
#>