3-D implementation of the Resnet + U-net deep learning architecture.

Creates a keras model of the U-net + ResNet deep learning architecture for image segmentation and regression with the paper available here:

createResUnetModel3D(inputImageSize, numberOfOutputs = 1,
  numberOfFiltersAtBaseLayer = 32, bottleNeckBlockDepthSchedule = c(3,
  4), convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2), dropoutRate = 0,
  weightDecay = 0.0001, mode = "classification")

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer.

bottleNeckBlockDepthSchedule
vector that provides the encoding layer schedule for the number of bottleneck blocks per long skip connection.

convolutionKernelSize
2-d vector defining the kernel size during the encoding path

deconvolutionKernelSize
2-d vector defining the kernel size during the decoding

dropoutRate
float between 0 and 1 to use between dense layers.

weightDecay
weighting parameter for L2 regularization of the kernel weights of the convolution layers. Default = 0.0.

mode
'classification' or 'regression'. Default = 'classification'.

Value

a res/u-net keras model

Details

https://arxiv.org/abs/1608.04117

This particular implementation was ported from the following python implementation:

https://github.com/veugene/fcn_maker/

Examples

library( ANTsRNet )
library( keras )

model <- createResUnetModel3D( c( 64, 64, 64, 1 ) )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_multilabel_dice_coefficient,
    metric_categorical_crossentropy ) )

print( model )

#> Model
#> Model: "model_17"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_14 (InputLayer)     (None, 64, 64, 64 0                                   
#> ________________________________________________________________________________
#> conv3d_79 (Conv3D)        (None, 64, 64, 64 896      input_14[0][0]             
#> ________________________________________________________________________________
#> batch_normalization_414 ( (None, 64, 64, 64 128      conv3d_79[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_39 (Thr (None, 64, 64, 64 0        batch_normalization_414[0][
#> ________________________________________________________________________________
#> max_pooling3d_5 (MaxPooli (None, 32, 32, 32 0        thresholded_re_lu_39[0][0] 
#> ________________________________________________________________________________
#> conv3d_81 (Conv3D)        (None, 32, 32, 32 1056     conv3d_79[0][0]            
#> ________________________________________________________________________________
#> conv3d_80 (Conv3D)        (None, 32, 32, 32 27680    max_pooling3d_5[0][0]      
#> ________________________________________________________________________________
#> add_44 (Add)              (None, 32, 32, 32 0        conv3d_81[0][0]            
#>                                                      conv3d_80[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_415 ( (None, 32, 32, 32 128      add_44[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_40 (Thr (None, 32, 32, 32 0        batch_normalization_415[0][
#> ________________________________________________________________________________
#> conv3d_82 (Conv3D)        (None, 16, 16, 16 1056     thresholded_re_lu_40[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_416 ( (None, 16, 16, 16 128      conv3d_82[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_41 (Thr (None, 16, 16, 16 0        batch_normalization_416[0][
#> ________________________________________________________________________________
#> conv3d_83 (Conv3D)        (None, 16, 16, 16 1056     thresholded_re_lu_41[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_417 ( (None, 16, 16, 16 128      conv3d_83[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_42 (Thr (None, 16, 16, 16 0        batch_normalization_417[0][
#> ________________________________________________________________________________
#> conv3d_85 (Conv3D)        (None, 16, 16, 16 4224     add_44[0][0]               
#> ________________________________________________________________________________
#> conv3d_84 (Conv3D)        (None, 16, 16, 16 4224     thresholded_re_lu_42[0][0] 
#> ________________________________________________________________________________
#> add_45 (Add)              (None, 16, 16, 16 0        conv3d_85[0][0]            
#>                                                      conv3d_84[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_418 ( (None, 16, 16, 16 512      add_45[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_43 (Thr (None, 16, 16, 16 0        batch_normalization_418[0][
#> ________________________________________________________________________________
#> conv3d_86 (Conv3D)        (None, 16, 16, 16 4128     thresholded_re_lu_43[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_419 ( (None, 16, 16, 16 128      conv3d_86[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_44 (Thr (None, 16, 16, 16 0        batch_normalization_419[0][
#> ________________________________________________________________________________
#> conv3d_88 (Conv3D)        (None, 16, 16, 16 16512    add_45[0][0]               
#> ________________________________________________________________________________
#> conv3d_87 (Conv3D)        (None, 16, 16, 16 4224     thresholded_re_lu_44[0][0] 
#> ________________________________________________________________________________
#> add_46 (Add)              (None, 16, 16, 16 0        conv3d_88[0][0]            
#>                                                      conv3d_87[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_420 ( (None, 16, 16, 16 512      add_46[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_45 (Thr (None, 16, 16, 16 0        batch_normalization_420[0][
#> ________________________________________________________________________________
#> conv3d_89 (Conv3D)        (None, 16, 16, 16 4128     thresholded_re_lu_45[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_421 ( (None, 16, 16, 16 128      conv3d_89[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_46 (Thr (None, 16, 16, 16 0        batch_normalization_421[0][
#> ________________________________________________________________________________
#> conv3d_91 (Conv3D)        (None, 16, 16, 16 16512    add_46[0][0]               
#> ________________________________________________________________________________
#> conv3d_90 (Conv3D)        (None, 16, 16, 16 4224     thresholded_re_lu_46[0][0] 
#> ________________________________________________________________________________
#> add_47 (Add)              (None, 16, 16, 16 0        conv3d_91[0][0]            
#>                                                      conv3d_90[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_422 ( (None, 16, 16, 16 512      add_47[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_47 (Thr (None, 16, 16, 16 0        batch_normalization_422[0][
#> ________________________________________________________________________________
#> conv3d_92 (Conv3D)        (None, 8, 8, 8, 6 8256     thresholded_re_lu_47[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_423 ( (None, 8, 8, 8, 6 256      conv3d_92[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_48 (Thr (None, 8, 8, 8, 6 0        batch_normalization_423[0][
#> ________________________________________________________________________________
#> conv3d_93 (Conv3D)        (None, 8, 8, 8, 6 4160     thresholded_re_lu_48[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_424 ( (None, 8, 8, 8, 6 256      conv3d_93[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_49 (Thr (None, 8, 8, 8, 6 0        batch_normalization_424[0][
#> ________________________________________________________________________________
#> conv3d_95 (Conv3D)        (None, 8, 8, 8, 2 33024    add_47[0][0]               
#> ________________________________________________________________________________
#> conv3d_94 (Conv3D)        (None, 8, 8, 8, 2 16640    thresholded_re_lu_49[0][0] 
#> ________________________________________________________________________________
#> add_48 (Add)              (None, 8, 8, 8, 2 0        conv3d_95[0][0]            
#>                                                      conv3d_94[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_425 ( (None, 8, 8, 8, 2 1024     add_48[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_50 (Thr (None, 8, 8, 8, 2 0        batch_normalization_425[0][
#> ________________________________________________________________________________
#> conv3d_96 (Conv3D)        (None, 8, 8, 8, 6 16448    thresholded_re_lu_50[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_426 ( (None, 8, 8, 8, 6 256      conv3d_96[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_51 (Thr (None, 8, 8, 8, 6 0        batch_normalization_426[0][
#> ________________________________________________________________________________
#> conv3d_98 (Conv3D)        (None, 8, 8, 8, 2 65792    add_48[0][0]               
#> ________________________________________________________________________________
#> conv3d_97 (Conv3D)        (None, 8, 8, 8, 2 16640    thresholded_re_lu_51[0][0] 
#> ________________________________________________________________________________
#> add_49 (Add)              (None, 8, 8, 8, 2 0        conv3d_98[0][0]            
#>                                                      conv3d_97[0][0]            
#> ________________________________________________________________________________
#> batch_normalization_427 ( (None, 8, 8, 8, 2 1024     add_49[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_52 (Thr (None, 8, 8, 8, 2 0        batch_normalization_427[0][
#> ________________________________________________________________________________
#> conv3d_99 (Conv3D)        (None, 8, 8, 8, 6 16448    thresholded_re_lu_52[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_428 ( (None, 8, 8, 8, 6 256      conv3d_99[0][0]            
#> ________________________________________________________________________________
#> thresholded_re_lu_53 (Thr (None, 8, 8, 8, 6 0        batch_normalization_428[0][
#> ________________________________________________________________________________
#> conv3d_101 (Conv3D)       (None, 8, 8, 8, 2 65792    add_49[0][0]               
#> ________________________________________________________________________________
#> conv3d_100 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_53[0][0] 
#> ________________________________________________________________________________
#> add_50 (Add)              (None, 8, 8, 8, 2 0        conv3d_101[0][0]           
#>                                                      conv3d_100[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_429 ( (None, 8, 8, 8, 2 1024     add_50[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_54 (Thr (None, 8, 8, 8, 2 0        batch_normalization_429[0][
#> ________________________________________________________________________________
#> conv3d_102 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_54[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_430 ( (None, 8, 8, 8, 6 256      conv3d_102[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_55 (Thr (None, 8, 8, 8, 6 0        batch_normalization_430[0][
#> ________________________________________________________________________________
#> conv3d_104 (Conv3D)       (None, 8, 8, 8, 2 65792    add_50[0][0]               
#> ________________________________________________________________________________
#> conv3d_103 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_55[0][0] 
#> ________________________________________________________________________________
#> add_51 (Add)              (None, 8, 8, 8, 2 0        conv3d_104[0][0]           
#>                                                      conv3d_103[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_431 ( (None, 8, 8, 8, 2 1024     add_51[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_56 (Thr (None, 8, 8, 8, 2 0        batch_normalization_431[0][
#> ________________________________________________________________________________
#> conv3d_105 (Conv3D)       (None, 4, 4, 4, 1 32896    thresholded_re_lu_56[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_432 ( (None, 4, 4, 4, 1 512      conv3d_105[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_57 (Thr (None, 4, 4, 4, 1 0        batch_normalization_432[0][
#> ________________________________________________________________________________
#> conv3d_106 (Conv3D)       (None, 4, 4, 4, 1 16512    thresholded_re_lu_57[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_433 ( (None, 4, 4, 4, 1 512      conv3d_106[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_58 (Thr (None, 4, 4, 4, 1 0        batch_normalization_433[0][
#> ________________________________________________________________________________
#> conv3d_108 (Conv3D)       (None, 4, 4, 4, 5 131584   add_51[0][0]               
#> ________________________________________________________________________________
#> conv3d_107 (Conv3D)       (None, 4, 4, 4, 5 66048    thresholded_re_lu_58[0][0] 
#> ________________________________________________________________________________
#> add_52 (Add)              (None, 4, 4, 4, 5 0        conv3d_108[0][0]           
#>                                                      conv3d_107[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_434 ( (None, 4, 4, 4, 5 2048     add_52[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_59 (Thr (None, 4, 4, 4, 5 0        batch_normalization_434[0][
#> ________________________________________________________________________________
#> conv3d_109 (Conv3D)       (None, 4, 4, 4, 1 65664    thresholded_re_lu_59[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_435 ( (None, 4, 4, 4, 1 512      conv3d_109[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_60 (Thr (None, 4, 4, 4, 1 0        batch_normalization_435[0][
#> ________________________________________________________________________________
#> conv3d_transpose_4 (Conv3 (None, 4, 4, 4, 1 131200   thresholded_re_lu_60[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_5 (Conv3 (None, 4, 4, 4, 5 262656   add_52[0][0]               
#> ________________________________________________________________________________
#> up_sampling3d_4 (UpSampli (None, 8, 8, 8, 1 0        conv3d_transpose_4[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_5 (UpSampli (None, 8, 8, 8, 5 0        conv3d_transpose_5[0][0]   
#> ________________________________________________________________________________
#> conv3d_110 (Conv3D)       (None, 8, 8, 8, 5 66048    up_sampling3d_4[0][0]      
#> ________________________________________________________________________________
#> add_53 (Add)              (None, 8, 8, 8, 5 0        up_sampling3d_5[0][0]      
#>                                                      conv3d_110[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_436 ( (None, 8, 8, 8, 5 2048     add_53[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_61 (Thr (None, 8, 8, 8, 5 0        batch_normalization_436[0][
#> ________________________________________________________________________________
#> conv3d_111 (Conv3D)       (None, 8, 8, 8, 6 32832    thresholded_re_lu_61[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_437 ( (None, 8, 8, 8, 6 256      conv3d_111[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_62 (Thr (None, 8, 8, 8, 6 0        batch_normalization_437[0][
#> ________________________________________________________________________________
#> conv3d_113 (Conv3D)       (None, 8, 8, 8, 2 131328   add_53[0][0]               
#> ________________________________________________________________________________
#> conv3d_112 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_62[0][0] 
#> ________________________________________________________________________________
#> add_54 (Add)              (None, 8, 8, 8, 2 0        conv3d_113[0][0]           
#>                                                      conv3d_112[0][0]           
#> ________________________________________________________________________________
#> conv3d_114 (Conv3D)       (None, 8, 8, 8, 2 65792    add_54[0][0]               
#> ________________________________________________________________________________
#> add_55 (Add)              (None, 8, 8, 8, 2 0        add_51[0][0]               
#>                                                      conv3d_114[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_438 ( (None, 8, 8, 8, 2 1024     add_55[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_63 (Thr (None, 8, 8, 8, 2 0        batch_normalization_438[0][
#> ________________________________________________________________________________
#> conv3d_115 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_63[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_439 ( (None, 8, 8, 8, 6 256      conv3d_115[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_64 (Thr (None, 8, 8, 8, 6 0        batch_normalization_439[0][
#> ________________________________________________________________________________
#> conv3d_117 (Conv3D)       (None, 8, 8, 8, 2 65792    add_55[0][0]               
#> ________________________________________________________________________________
#> conv3d_116 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_64[0][0] 
#> ________________________________________________________________________________
#> add_56 (Add)              (None, 8, 8, 8, 2 0        conv3d_117[0][0]           
#>                                                      conv3d_116[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_440 ( (None, 8, 8, 8, 2 1024     add_56[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_65 (Thr (None, 8, 8, 8, 2 0        batch_normalization_440[0][
#> ________________________________________________________________________________
#> conv3d_118 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_65[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_441 ( (None, 8, 8, 8, 6 256      conv3d_118[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_66 (Thr (None, 8, 8, 8, 6 0        batch_normalization_441[0][
#> ________________________________________________________________________________
#> conv3d_120 (Conv3D)       (None, 8, 8, 8, 2 65792    add_56[0][0]               
#> ________________________________________________________________________________
#> conv3d_119 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_66[0][0] 
#> ________________________________________________________________________________
#> add_57 (Add)              (None, 8, 8, 8, 2 0        conv3d_120[0][0]           
#>                                                      conv3d_119[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_442 ( (None, 8, 8, 8, 2 1024     add_57[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_67 (Thr (None, 8, 8, 8, 2 0        batch_normalization_442[0][
#> ________________________________________________________________________________
#> conv3d_121 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_67[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_443 ( (None, 8, 8, 8, 6 256      conv3d_121[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_68 (Thr (None, 8, 8, 8, 6 0        batch_normalization_443[0][
#> ________________________________________________________________________________
#> conv3d_transpose_6 (Conv3 (None, 8, 8, 8, 6 32832    thresholded_re_lu_68[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_7 (Conv3 (None, 8, 8, 8, 2 65792    add_57[0][0]               
#> ________________________________________________________________________________
#> up_sampling3d_6 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_6[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_7 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_7[0][0]   
#> ________________________________________________________________________________
#> conv3d_122 (Conv3D)       (None, 16, 16, 16 16640    up_sampling3d_6[0][0]      
#> ________________________________________________________________________________
#> add_58 (Add)              (None, 16, 16, 16 0        up_sampling3d_7[0][0]      
#>                                                      conv3d_122[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_444 ( (None, 16, 16, 16 1024     add_58[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_69 (Thr (None, 16, 16, 16 0        batch_normalization_444[0][
#> ________________________________________________________________________________
#> conv3d_123 (Conv3D)       (None, 16, 16, 16 8224     thresholded_re_lu_69[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_445 ( (None, 16, 16, 16 128      conv3d_123[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_70 (Thr (None, 16, 16, 16 0        batch_normalization_445[0][
#> ________________________________________________________________________________
#> conv3d_125 (Conv3D)       (None, 16, 16, 16 32896    add_58[0][0]               
#> ________________________________________________________________________________
#> conv3d_124 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_70[0][0] 
#> ________________________________________________________________________________
#> add_59 (Add)              (None, 16, 16, 16 0        conv3d_125[0][0]           
#>                                                      conv3d_124[0][0]           
#> ________________________________________________________________________________
#> conv3d_126 (Conv3D)       (None, 16, 16, 16 16512    add_59[0][0]               
#> ________________________________________________________________________________
#> add_60 (Add)              (None, 16, 16, 16 0        add_47[0][0]               
#>                                                      conv3d_126[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_446 ( (None, 16, 16, 16 512      add_60[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_71 (Thr (None, 16, 16, 16 0        batch_normalization_446[0][
#> ________________________________________________________________________________
#> conv3d_127 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_71[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_447 ( (None, 16, 16, 16 128      conv3d_127[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_72 (Thr (None, 16, 16, 16 0        batch_normalization_447[0][
#> ________________________________________________________________________________
#> conv3d_129 (Conv3D)       (None, 16, 16, 16 16512    add_60[0][0]               
#> ________________________________________________________________________________
#> conv3d_128 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_72[0][0] 
#> ________________________________________________________________________________
#> add_61 (Add)              (None, 16, 16, 16 0        conv3d_129[0][0]           
#>                                                      conv3d_128[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_448 ( (None, 16, 16, 16 512      add_61[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_73 (Thr (None, 16, 16, 16 0        batch_normalization_448[0][
#> ________________________________________________________________________________
#> conv3d_130 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_73[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_449 ( (None, 16, 16, 16 128      conv3d_130[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_74 (Thr (None, 16, 16, 16 0        batch_normalization_449[0][
#> ________________________________________________________________________________
#> conv3d_transpose_8 (Conv3 (None, 16, 16, 16 8224     thresholded_re_lu_74[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_9 (Conv3 (None, 16, 16, 16 16512    add_61[0][0]               
#> ________________________________________________________________________________
#> up_sampling3d_8 (UpSampli (None, 32, 32, 32 0        conv3d_transpose_8[0][0]   
#> ________________________________________________________________________________
#> up_sampling3d_9 (UpSampli (None, 32, 32, 32 0        conv3d_transpose_9[0][0]   
#> ________________________________________________________________________________
#> conv3d_131 (Conv3D)       (None, 32, 32, 32 4224     up_sampling3d_8[0][0]      
#> ________________________________________________________________________________
#> add_62 (Add)              (None, 32, 32, 32 0        up_sampling3d_9[0][0]      
#>                                                      conv3d_131[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_450 ( (None, 32, 32, 32 512      add_62[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_75 (Thr (None, 32, 32, 32 0        batch_normalization_450[0][
#> ________________________________________________________________________________
#> conv3d_132 (Conv3D)       (None, 32, 32, 32 110624   thresholded_re_lu_75[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_11 (Conv (None, 32, 32, 32 4128     add_62[0][0]               
#> ________________________________________________________________________________
#> conv3d_transpose_10 (Conv (None, 32, 32, 32 8224     conv3d_132[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_11 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_11[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_10 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_10[0][0]  
#> ________________________________________________________________________________
#> add_63 (Add)              (None, 64, 64, 64 0        up_sampling3d_11[0][0]     
#>                                                      up_sampling3d_10[0][0]     
#> ________________________________________________________________________________
#> conv3d_133 (Conv3D)       (None, 64, 64, 64 27680    add_63[0][0]               
#> ________________________________________________________________________________
#> add_64 (Add)              (None, 64, 64, 64 0        conv3d_79[0][0]            
#>                                                      conv3d_133[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_451 ( (None, 64, 64, 64 128      add_64[0][0]               
#> ________________________________________________________________________________
#> thresholded_re_lu_76 (Thr (None, 64, 64, 64 0        batch_normalization_451[0][
#> ________________________________________________________________________________
#> conv3d_134 (Conv3D)       (None, 64, 64, 64 33       thresholded_re_lu_76[0][0] 
#> ================================================================================
#> Total params: 2,088,289
#> Trainable params: 2,078,049
#> Non-trainable params: 10,240
#> ________________________________________________________________________________
#> 
#>