3-D implementation of the Resnet + U-net deep learning architecture.

Creates a keras model of the U-net + ResNet deep learning architecture for image segmentation and regression with the paper available here:

createResUnetModel3D(
  inputImageSize,
  numberOfOutputs = 1,
  numberOfFiltersAtBaseLayer = 32,
  bottleNeckBlockDepthSchedule = c(3, 4),
  convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2),
  dropoutRate = 0,
  weightDecay = 0.0001,
  mode = c("classification", "regression")
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer.

bottleNeckBlockDepthSchedule
vector that provides the encoding layer schedule for the number of bottleneck blocks per long skip connection.

convolutionKernelSize
2-d vector defining the kernel size during the encoding path

deconvolutionKernelSize
2-d vector defining the kernel size during the decoding

dropoutRate
float between 0 and 1 to use between dense layers.

weightDecay
weighting parameter for L2 regularization of the kernel weights of the convolution layers. Default = 0.0.

mode
'classification' or 'regression'.

Value

a res/u-net keras model

Details

    \url{https://arxiv.org/abs/1608.04117}

This particular implementation was ported from the following python implementation:

    \url{https://github.com/veugene/fcn_maker/}

Author

Tustison NJ

Examples

library( ANTsRNet )
library( keras )

model <- createResUnetModel3D( c( 64, 64, 64, 1 ) )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_multilabel_dice_coefficient,
    metric_categorical_crossentropy ) )

print( model )

#> Model
#> Model: "model_35"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_32 (InputLayer)     [(None, 64, 64, 6 0                                   
#> ________________________________________________________________________________
#> conv3d_144 (Conv3D)       (None, 64, 64, 64 896      input_32[0][0]             
#> ________________________________________________________________________________
#> batch_normalization_692 ( (None, 64, 64, 64 128      conv3d_144[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_38 (Thr (None, 64, 64, 64 0        batch_normalization_692[0][
#> ________________________________________________________________________________
#> max_pooling3d_4 (MaxPooli (None, 32, 32, 32 0        thresholded_re_lu_38[0][0] 
#> ________________________________________________________________________________
#> conv3d_146 (Conv3D)       (None, 32, 32, 32 1056     conv3d_144[0][0]           
#> ________________________________________________________________________________
#> conv3d_145 (Conv3D)       (None, 32, 32, 32 27680    max_pooling3d_4[0][0]      
#> ________________________________________________________________________________
#> add_190 (Add)             (None, 32, 32, 32 0        conv3d_146[0][0]           
#>                                                      conv3d_145[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_693 ( (None, 32, 32, 32 128      add_190[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_39 (Thr (None, 32, 32, 32 0        batch_normalization_693[0][
#> ________________________________________________________________________________
#> conv3d_147 (Conv3D)       (None, 16, 16, 16 1056     thresholded_re_lu_39[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_694 ( (None, 16, 16, 16 128      conv3d_147[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_40 (Thr (None, 16, 16, 16 0        batch_normalization_694[0][
#> ________________________________________________________________________________
#> conv3d_148 (Conv3D)       (None, 16, 16, 16 1056     thresholded_re_lu_40[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_695 ( (None, 16, 16, 16 128      conv3d_148[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_41 (Thr (None, 16, 16, 16 0        batch_normalization_695[0][
#> ________________________________________________________________________________
#> conv3d_150 (Conv3D)       (None, 16, 16, 16 4224     add_190[0][0]              
#> ________________________________________________________________________________
#> conv3d_149 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_41[0][0] 
#> ________________________________________________________________________________
#> add_191 (Add)             (None, 16, 16, 16 0        conv3d_150[0][0]           
#>                                                      conv3d_149[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_696 ( (None, 16, 16, 16 512      add_191[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_42 (Thr (None, 16, 16, 16 0        batch_normalization_696[0][
#> ________________________________________________________________________________
#> conv3d_151 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_42[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_697 ( (None, 16, 16, 16 128      conv3d_151[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_43 (Thr (None, 16, 16, 16 0        batch_normalization_697[0][
#> ________________________________________________________________________________
#> conv3d_153 (Conv3D)       (None, 16, 16, 16 16512    add_191[0][0]              
#> ________________________________________________________________________________
#> conv3d_152 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_43[0][0] 
#> ________________________________________________________________________________
#> add_192 (Add)             (None, 16, 16, 16 0        conv3d_153[0][0]           
#>                                                      conv3d_152[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_698 ( (None, 16, 16, 16 512      add_192[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_44 (Thr (None, 16, 16, 16 0        batch_normalization_698[0][
#> ________________________________________________________________________________
#> conv3d_154 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_44[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_699 ( (None, 16, 16, 16 128      conv3d_154[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_45 (Thr (None, 16, 16, 16 0        batch_normalization_699[0][
#> ________________________________________________________________________________
#> conv3d_156 (Conv3D)       (None, 16, 16, 16 16512    add_192[0][0]              
#> ________________________________________________________________________________
#> conv3d_155 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_45[0][0] 
#> ________________________________________________________________________________
#> add_193 (Add)             (None, 16, 16, 16 0        conv3d_156[0][0]           
#>                                                      conv3d_155[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_700 ( (None, 16, 16, 16 512      add_193[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_46 (Thr (None, 16, 16, 16 0        batch_normalization_700[0][
#> ________________________________________________________________________________
#> conv3d_157 (Conv3D)       (None, 8, 8, 8, 6 8256     thresholded_re_lu_46[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_701 ( (None, 8, 8, 8, 6 256      conv3d_157[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_47 (Thr (None, 8, 8, 8, 6 0        batch_normalization_701[0][
#> ________________________________________________________________________________
#> conv3d_158 (Conv3D)       (None, 8, 8, 8, 6 4160     thresholded_re_lu_47[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_702 ( (None, 8, 8, 8, 6 256      conv3d_158[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_48 (Thr (None, 8, 8, 8, 6 0        batch_normalization_702[0][
#> ________________________________________________________________________________
#> conv3d_160 (Conv3D)       (None, 8, 8, 8, 2 33024    add_193[0][0]              
#> ________________________________________________________________________________
#> conv3d_159 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_48[0][0] 
#> ________________________________________________________________________________
#> add_194 (Add)             (None, 8, 8, 8, 2 0        conv3d_160[0][0]           
#>                                                      conv3d_159[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_703 ( (None, 8, 8, 8, 2 1024     add_194[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_49 (Thr (None, 8, 8, 8, 2 0        batch_normalization_703[0][
#> ________________________________________________________________________________
#> conv3d_161 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_49[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_704 ( (None, 8, 8, 8, 6 256      conv3d_161[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_50 (Thr (None, 8, 8, 8, 6 0        batch_normalization_704[0][
#> ________________________________________________________________________________
#> conv3d_163 (Conv3D)       (None, 8, 8, 8, 2 65792    add_194[0][0]              
#> ________________________________________________________________________________
#> conv3d_162 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_50[0][0] 
#> ________________________________________________________________________________
#> add_195 (Add)             (None, 8, 8, 8, 2 0        conv3d_163[0][0]           
#>                                                      conv3d_162[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_705 ( (None, 8, 8, 8, 2 1024     add_195[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_51 (Thr (None, 8, 8, 8, 2 0        batch_normalization_705[0][
#> ________________________________________________________________________________
#> conv3d_164 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_51[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_706 ( (None, 8, 8, 8, 6 256      conv3d_164[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_52 (Thr (None, 8, 8, 8, 6 0        batch_normalization_706[0][
#> ________________________________________________________________________________
#> conv3d_166 (Conv3D)       (None, 8, 8, 8, 2 65792    add_195[0][0]              
#> ________________________________________________________________________________
#> conv3d_165 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_52[0][0] 
#> ________________________________________________________________________________
#> add_196 (Add)             (None, 8, 8, 8, 2 0        conv3d_166[0][0]           
#>                                                      conv3d_165[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_707 ( (None, 8, 8, 8, 2 1024     add_196[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_53 (Thr (None, 8, 8, 8, 2 0        batch_normalization_707[0][
#> ________________________________________________________________________________
#> conv3d_167 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_53[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_708 ( (None, 8, 8, 8, 6 256      conv3d_167[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_54 (Thr (None, 8, 8, 8, 6 0        batch_normalization_708[0][
#> ________________________________________________________________________________
#> conv3d_169 (Conv3D)       (None, 8, 8, 8, 2 65792    add_196[0][0]              
#> ________________________________________________________________________________
#> conv3d_168 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_54[0][0] 
#> ________________________________________________________________________________
#> add_197 (Add)             (None, 8, 8, 8, 2 0        conv3d_169[0][0]           
#>                                                      conv3d_168[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_709 ( (None, 8, 8, 8, 2 1024     add_197[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_55 (Thr (None, 8, 8, 8, 2 0        batch_normalization_709[0][
#> ________________________________________________________________________________
#> conv3d_170 (Conv3D)       (None, 4, 4, 4, 1 32896    thresholded_re_lu_55[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_710 ( (None, 4, 4, 4, 1 512      conv3d_170[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_56 (Thr (None, 4, 4, 4, 1 0        batch_normalization_710[0][
#> ________________________________________________________________________________
#> conv3d_171 (Conv3D)       (None, 4, 4, 4, 1 16512    thresholded_re_lu_56[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_711 ( (None, 4, 4, 4, 1 512      conv3d_171[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_57 (Thr (None, 4, 4, 4, 1 0        batch_normalization_711[0][
#> ________________________________________________________________________________
#> conv3d_173 (Conv3D)       (None, 4, 4, 4, 5 131584   add_197[0][0]              
#> ________________________________________________________________________________
#> conv3d_172 (Conv3D)       (None, 4, 4, 4, 5 66048    thresholded_re_lu_57[0][0] 
#> ________________________________________________________________________________
#> add_198 (Add)             (None, 4, 4, 4, 5 0        conv3d_173[0][0]           
#>                                                      conv3d_172[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_712 ( (None, 4, 4, 4, 5 2048     add_198[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_58 (Thr (None, 4, 4, 4, 5 0        batch_normalization_712[0][
#> ________________________________________________________________________________
#> conv3d_174 (Conv3D)       (None, 4, 4, 4, 1 65664    thresholded_re_lu_58[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_713 ( (None, 4, 4, 4, 1 512      conv3d_174[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_59 (Thr (None, 4, 4, 4, 1 0        batch_normalization_713[0][
#> ________________________________________________________________________________
#> conv3d_transpose_28 (Conv (None, 4, 4, 4, 1 131200   thresholded_re_lu_59[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_29 (Conv (None, 4, 4, 4, 5 262656   add_198[0][0]              
#> ________________________________________________________________________________
#> up_sampling3d_4 (UpSampli (None, 8, 8, 8, 1 0        conv3d_transpose_28[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_5 (UpSampli (None, 8, 8, 8, 5 0        conv3d_transpose_29[0][0]  
#> ________________________________________________________________________________
#> conv3d_175 (Conv3D)       (None, 8, 8, 8, 5 66048    up_sampling3d_4[0][0]      
#> ________________________________________________________________________________
#> add_199 (Add)             (None, 8, 8, 8, 5 0        up_sampling3d_5[0][0]      
#>                                                      conv3d_175[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_714 ( (None, 8, 8, 8, 5 2048     add_199[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_60 (Thr (None, 8, 8, 8, 5 0        batch_normalization_714[0][
#> ________________________________________________________________________________
#> conv3d_176 (Conv3D)       (None, 8, 8, 8, 6 32832    thresholded_re_lu_60[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_715 ( (None, 8, 8, 8, 6 256      conv3d_176[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_61 (Thr (None, 8, 8, 8, 6 0        batch_normalization_715[0][
#> ________________________________________________________________________________
#> conv3d_178 (Conv3D)       (None, 8, 8, 8, 2 131328   add_199[0][0]              
#> ________________________________________________________________________________
#> conv3d_177 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_61[0][0] 
#> ________________________________________________________________________________
#> add_200 (Add)             (None, 8, 8, 8, 2 0        conv3d_178[0][0]           
#>                                                      conv3d_177[0][0]           
#> ________________________________________________________________________________
#> conv3d_179 (Conv3D)       (None, 8, 8, 8, 2 65792    add_200[0][0]              
#> ________________________________________________________________________________
#> add_201 (Add)             (None, 8, 8, 8, 2 0        add_197[0][0]              
#>                                                      conv3d_179[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_716 ( (None, 8, 8, 8, 2 1024     add_201[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_62 (Thr (None, 8, 8, 8, 2 0        batch_normalization_716[0][
#> ________________________________________________________________________________
#> conv3d_180 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_62[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_717 ( (None, 8, 8, 8, 6 256      conv3d_180[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_63 (Thr (None, 8, 8, 8, 6 0        batch_normalization_717[0][
#> ________________________________________________________________________________
#> conv3d_182 (Conv3D)       (None, 8, 8, 8, 2 65792    add_201[0][0]              
#> ________________________________________________________________________________
#> conv3d_181 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_63[0][0] 
#> ________________________________________________________________________________
#> add_202 (Add)             (None, 8, 8, 8, 2 0        conv3d_182[0][0]           
#>                                                      conv3d_181[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_718 ( (None, 8, 8, 8, 2 1024     add_202[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_64 (Thr (None, 8, 8, 8, 2 0        batch_normalization_718[0][
#> ________________________________________________________________________________
#> conv3d_183 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_64[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_719 ( (None, 8, 8, 8, 6 256      conv3d_183[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_65 (Thr (None, 8, 8, 8, 6 0        batch_normalization_719[0][
#> ________________________________________________________________________________
#> conv3d_185 (Conv3D)       (None, 8, 8, 8, 2 65792    add_202[0][0]              
#> ________________________________________________________________________________
#> conv3d_184 (Conv3D)       (None, 8, 8, 8, 2 16640    thresholded_re_lu_65[0][0] 
#> ________________________________________________________________________________
#> add_203 (Add)             (None, 8, 8, 8, 2 0        conv3d_185[0][0]           
#>                                                      conv3d_184[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_720 ( (None, 8, 8, 8, 2 1024     add_203[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_66 (Thr (None, 8, 8, 8, 2 0        batch_normalization_720[0][
#> ________________________________________________________________________________
#> conv3d_186 (Conv3D)       (None, 8, 8, 8, 6 16448    thresholded_re_lu_66[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_721 ( (None, 8, 8, 8, 6 256      conv3d_186[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_67 (Thr (None, 8, 8, 8, 6 0        batch_normalization_721[0][
#> ________________________________________________________________________________
#> conv3d_transpose_30 (Conv (None, 8, 8, 8, 6 32832    thresholded_re_lu_67[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_31 (Conv (None, 8, 8, 8, 2 65792    add_203[0][0]              
#> ________________________________________________________________________________
#> up_sampling3d_6 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_30[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_7 (UpSampli (None, 16, 16, 16 0        conv3d_transpose_31[0][0]  
#> ________________________________________________________________________________
#> conv3d_187 (Conv3D)       (None, 16, 16, 16 16640    up_sampling3d_6[0][0]      
#> ________________________________________________________________________________
#> add_204 (Add)             (None, 16, 16, 16 0        up_sampling3d_7[0][0]      
#>                                                      conv3d_187[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_722 ( (None, 16, 16, 16 1024     add_204[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_68 (Thr (None, 16, 16, 16 0        batch_normalization_722[0][
#> ________________________________________________________________________________
#> conv3d_188 (Conv3D)       (None, 16, 16, 16 8224     thresholded_re_lu_68[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_723 ( (None, 16, 16, 16 128      conv3d_188[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_69 (Thr (None, 16, 16, 16 0        batch_normalization_723[0][
#> ________________________________________________________________________________
#> conv3d_190 (Conv3D)       (None, 16, 16, 16 32896    add_204[0][0]              
#> ________________________________________________________________________________
#> conv3d_189 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_69[0][0] 
#> ________________________________________________________________________________
#> add_205 (Add)             (None, 16, 16, 16 0        conv3d_190[0][0]           
#>                                                      conv3d_189[0][0]           
#> ________________________________________________________________________________
#> conv3d_191 (Conv3D)       (None, 16, 16, 16 16512    add_205[0][0]              
#> ________________________________________________________________________________
#> add_206 (Add)             (None, 16, 16, 16 0        add_193[0][0]              
#>                                                      conv3d_191[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_724 ( (None, 16, 16, 16 512      add_206[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_70 (Thr (None, 16, 16, 16 0        batch_normalization_724[0][
#> ________________________________________________________________________________
#> conv3d_192 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_70[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_725 ( (None, 16, 16, 16 128      conv3d_192[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_71 (Thr (None, 16, 16, 16 0        batch_normalization_725[0][
#> ________________________________________________________________________________
#> conv3d_194 (Conv3D)       (None, 16, 16, 16 16512    add_206[0][0]              
#> ________________________________________________________________________________
#> conv3d_193 (Conv3D)       (None, 16, 16, 16 4224     thresholded_re_lu_71[0][0] 
#> ________________________________________________________________________________
#> add_207 (Add)             (None, 16, 16, 16 0        conv3d_194[0][0]           
#>                                                      conv3d_193[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_726 ( (None, 16, 16, 16 512      add_207[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_72 (Thr (None, 16, 16, 16 0        batch_normalization_726[0][
#> ________________________________________________________________________________
#> conv3d_195 (Conv3D)       (None, 16, 16, 16 4128     thresholded_re_lu_72[0][0] 
#> ________________________________________________________________________________
#> batch_normalization_727 ( (None, 16, 16, 16 128      conv3d_195[0][0]           
#> ________________________________________________________________________________
#> thresholded_re_lu_73 (Thr (None, 16, 16, 16 0        batch_normalization_727[0][
#> ________________________________________________________________________________
#> conv3d_transpose_32 (Conv (None, 16, 16, 16 8224     thresholded_re_lu_73[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_33 (Conv (None, 16, 16, 16 16512    add_207[0][0]              
#> ________________________________________________________________________________
#> up_sampling3d_8 (UpSampli (None, 32, 32, 32 0        conv3d_transpose_32[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_9 (UpSampli (None, 32, 32, 32 0        conv3d_transpose_33[0][0]  
#> ________________________________________________________________________________
#> conv3d_196 (Conv3D)       (None, 32, 32, 32 4224     up_sampling3d_8[0][0]      
#> ________________________________________________________________________________
#> add_208 (Add)             (None, 32, 32, 32 0        up_sampling3d_9[0][0]      
#>                                                      conv3d_196[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_728 ( (None, 32, 32, 32 512      add_208[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_74 (Thr (None, 32, 32, 32 0        batch_normalization_728[0][
#> ________________________________________________________________________________
#> conv3d_197 (Conv3D)       (None, 32, 32, 32 110624   thresholded_re_lu_74[0][0] 
#> ________________________________________________________________________________
#> conv3d_transpose_35 (Conv (None, 32, 32, 32 4128     add_208[0][0]              
#> ________________________________________________________________________________
#> conv3d_transpose_34 (Conv (None, 32, 32, 32 8224     conv3d_197[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_11 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_35[0][0]  
#> ________________________________________________________________________________
#> up_sampling3d_10 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_34[0][0]  
#> ________________________________________________________________________________
#> add_209 (Add)             (None, 64, 64, 64 0        up_sampling3d_11[0][0]     
#>                                                      up_sampling3d_10[0][0]     
#> ________________________________________________________________________________
#> conv3d_198 (Conv3D)       (None, 64, 64, 64 27680    add_209[0][0]              
#> ________________________________________________________________________________
#> add_210 (Add)             (None, 64, 64, 64 0        conv3d_144[0][0]           
#>                                                      conv3d_198[0][0]           
#> ________________________________________________________________________________
#> batch_normalization_729 ( (None, 64, 64, 64 128      add_210[0][0]              
#> ________________________________________________________________________________
#> thresholded_re_lu_75 (Thr (None, 64, 64, 64 0        batch_normalization_729[0][
#> ________________________________________________________________________________
#> conv3d_199 (Conv3D)       (None, 64, 64, 64 33       thresholded_re_lu_75[0][0] 
#> ================================================================================
#> Total params: 2,088,289
#> Trainable params: 2,078,049
#> Non-trainable params: 10,240
#> ________________________________________________________________________________
#> 
#>