3-D image segmentation implementation of the U-net deep learning architecture.

Creates a keras model of the U-net deep learning architecture for image segmentation and regression. More information is provided at the authors' website:

createUnetModel3D(inputImageSize, numberOfOutputs = 2,
  numberOfLayers = 4, numberOfFiltersAtBaseLayer = 32,
  convolutionKernelSize = c(3, 3, 3), deconvolutionKernelSize = c(2, 2,
  2), poolSize = c(2, 2, 2), strides = c(2, 2, 2), dropoutRate = 0,
  weightDecay = 0, mode = "classification")

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

numberOfLayers
number of encoding/decoding layers.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer.

convolutionKernelSize
3-d vector defining the kernel size during the encoding path.

deconvolutionKernelSize
3-d vector defining the kernel size during the decoding.

poolSize
3-d vector defining the region for each pooling layer.

strides
3-d vector describing the stride length in each direction.

dropoutRate
float between 0 and 1 to use between dense layers.

weightDecay
weighting parameter for L2 regularization of the kernel weights of the convolution layers. Default = 0.0.

mode
'classification' or 'regression'. Default = 'classification'.

Value

a u-net keras model

Details

https://lmb.informatik.uni-freiburg.de/people/ronneber/u-net/

with the paper available here:

https://arxiv.org/abs/1505.04597

This particular implementation was influenced by the following python implementation:

https://github.com/joelthelion/ultrasound-nerve-segmentation

Examples

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )
library( keras )

model <- createUnetModel3D( c( 64, 64, 64, 1 ) )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = metric_multilabel_dice_coefficient )

print( model )

#> Model
#> Model: "model_18"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_15 (InputLayer)     (None, 64, 64, 64 0                                   
#> ________________________________________________________________________________
#> conv3d_120 (Conv3D)       (None, 64, 64, 64 896      input_15[0][0]             
#> ________________________________________________________________________________
#> conv3d_121 (Conv3D)       (None, 64, 64, 64 27680    conv3d_120[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_3 (MaxPooli (None, 32, 32, 32 0        conv3d_121[0][0]           
#> ________________________________________________________________________________
#> conv3d_122 (Conv3D)       (None, 32, 32, 32 55360    max_pooling3d_3[0][0]      
#> ________________________________________________________________________________
#> conv3d_123 (Conv3D)       (None, 32, 32, 32 110656   conv3d_122[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_4 (MaxPooli (None, 16, 16, 16 0        conv3d_123[0][0]           
#> ________________________________________________________________________________
#> conv3d_124 (Conv3D)       (None, 16, 16, 16 221312   max_pooling3d_4[0][0]      
#> ________________________________________________________________________________
#> conv3d_125 (Conv3D)       (None, 16, 16, 16 442496   conv3d_124[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_5 (MaxPooli (None, 8, 8, 8, 1 0        conv3d_125[0][0]           
#> ________________________________________________________________________________
#> conv3d_126 (Conv3D)       (None, 8, 8, 8, 2 884992   max_pooling3d_5[0][0]      
#> ________________________________________________________________________________
#> conv3d_127 (Conv3D)       (None, 8, 8, 8, 2 1769728  conv3d_126[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_9 (Conv3 (None, 8, 8, 8, 1 262272   conv3d_127[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_12 (UpSampl (None, 16, 16, 16 0        conv3d_transpose_9[0][0]   
#> ________________________________________________________________________________
#> concatenate_134 (Concaten (None, 16, 16, 16 0        up_sampling3d_12[0][0]     
#>                                                      conv3d_125[0][0]           
#> ________________________________________________________________________________
#> conv3d_128 (Conv3D)       (None, 16, 16, 16 884864   concatenate_134[0][0]      
#> ________________________________________________________________________________
#> conv3d_129 (Conv3D)       (None, 16, 16, 16 442496   conv3d_128[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_10 (Conv (None, 16, 16, 16 65600    conv3d_129[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_13 (UpSampl (None, 32, 32, 32 0        conv3d_transpose_10[0][0]  
#> ________________________________________________________________________________
#> concatenate_135 (Concaten (None, 32, 32, 32 0        up_sampling3d_13[0][0]     
#>                                                      conv3d_123[0][0]           
#> ________________________________________________________________________________
#> conv3d_130 (Conv3D)       (None, 32, 32, 32 221248   concatenate_135[0][0]      
#> ________________________________________________________________________________
#> conv3d_131 (Conv3D)       (None, 32, 32, 32 110656   conv3d_130[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_11 (Conv (None, 32, 32, 32 16416    conv3d_131[0][0]           
#> ________________________________________________________________________________
#> up_sampling3d_14 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_11[0][0]  
#> ________________________________________________________________________________
#> concatenate_136 (Concaten (None, 64, 64, 64 0        up_sampling3d_14[0][0]     
#>                                                      conv3d_121[0][0]           
#> ________________________________________________________________________________
#> conv3d_132 (Conv3D)       (None, 64, 64, 64 55328    concatenate_136[0][0]      
#> ________________________________________________________________________________
#> conv3d_133 (Conv3D)       (None, 64, 64, 64 27680    conv3d_132[0][0]           
#> ________________________________________________________________________________
#> conv3d_134 (Conv3D)       (None, 64, 64, 64 66       conv3d_133[0][0]           
#> ================================================================================
#> Total params: 5,599,746
#> Trainable params: 5,599,746
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> 
#>