3-D image segmentation implementation of the U-net deep learning architecture.

Creates a keras model of the U-net deep learning architecture for image segmentation and regression. More information is provided at the authors' website:

createUnetModel3D(
  inputImageSize,
  numberOfOutputs = 2,
  scalarOutputSize = 0,
  scalarOutputActivation = "relu",
  numberOfLayers = 4,
  numberOfFiltersAtBaseLayer = 32,
  numberOfFilters = NULL,
  convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2),
  poolSize = c(2, 2, 2),
  strides = c(2, 2, 2),
  dropoutRate = 0,
  weightDecay = 0,
  mode = c("classification", "regression", "sigmoid"),
  additionalOptions = NA
)

Arguments

inputImageSize
Used for specifying the input tensor shape. The shape (or dimension) of that tensor is the image dimensions followed by the number of channels (e.g., red, green, and blue). The batch size (i.e., number of training images) is not specified a priori.

numberOfOutputs
Meaning depends on the mode. For 'classification' this is the number of segmentation labels. For 'regression' this is the number of outputs.

scalarOutputSize
if greater than 0, a global average pooling from each encoding layer is concatenated to a dense layer as a secondary output.

scalarOutputActivation
activation for nonzero output scalar.

numberOfLayers
number of encoding/decoding layers.

numberOfFiltersAtBaseLayer
number of filters at the beginning and end of the 'U'. Doubles at each descending/ascending layer.

numberOfFilters
vector explicitly setting the number of filters at each layer. One can either set this or numberOfLayers and

convolutionKernelSize
3-d vector defining the kernel size during the encoding path.

deconvolutionKernelSize
3-d vector defining the kernel size during the decoding.

poolSize
3-d vector defining the region for each pooling layer.

strides
3-d vector describing the stride length in each direction.

dropoutRate
float between 0 and 1 to use between dense layers.

weightDecay
weighting parameter for L2 regularization of the kernel weights of the convolution layers. Default = 0.0.

mode
'classification' or 'regression' or 'sigmoid'.

additionalOptions
string or vector of strings specifying specific configuration add-ons/tweaks:

"attentionGating"attention-unet variant in https://pubmed.ncbi.nlm.nih.gov/33288961/

"nnUnetActivationStyle"U-net activation explained in https://pubmed.ncbi.nlm.nih.gov/33288961/

"initialConvolutionalKernelSizeX"Set the first two convolutional layer kernel sizes to X.

Value

a u-net keras model

Details

    \url{https://lmb.informatik.uni-freiburg.de/people/ronneber/u-net/}

with the paper available here:

    \url{https://arxiv.org/abs/1505.04597}

This particular implementation was influenced by the following python implementation:

    \url{https://github.com/joelthelion/ultrasound-nerve-segmentation}

Author

Tustison NJ

Examples

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )
library( keras )

model <- createUnetModel3D( c( 64, 64, 64, 1 ) )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

loss_dice <- function( y_true, y_pred ) {
  -multilabel_dice_coefficient(y_true, y_pred)
}
attr(loss_dice, "py_function_name") <- "multilabel_dice_coefficient"

model %>% compile( loss = loss_dice,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = metric_multilabel_dice_coefficient )

#> Error in py_call_impl(callable, dots$args, dots$keywords): TypeError: Failed to convert object of type <class 'function'> to Tensor. Contents: <function make_python_function.<locals>.python_function at 0x7fa47023cd08>. Consider casting elements to a supported type.
#> 
#> Detailed traceback:
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/training/tracking/base.py", line 517, in _method_wrapper
#>     result = method(self, *args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/engine/training_v1.py", line 450, in compile
#>     masks=self._prepare_output_masks())
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/engine/training_v1.py", line 1957, in _handle_metrics
#>     target, output, output_mask))
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/engine/training_v1.py", line 1908, in _handle_per_output_metrics
#>     metric_fn, y_true, y_pred, weights=weights, mask=mask)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/engine/training_utils_v1.py", line 1064, in call_metric_function
#>     return metric_fn(y_true, y_pred, sample_weight=weights)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/metrics.py", line 237, in __call__
#>     replica_local_fn, *args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/distribute/distributed_training_utils.py", line 66, in call_replica_local_fn
#>     return fn(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/metrics.py", line 217, in replica_local_fn
#>     update_op = self.update_state(*args, **kwargs)  # pylint: disable=not-callable
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/utils/metrics_utils.py", line 90, in decorated
#>     update_op = update_state_fn(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/metrics.py", line 177, in update_state_fn
#>     return ag_update_state(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/autograph/impl/api.py", line 667, in wrapper
#>     return converted_call(f, args, kwargs, options=options)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/autograph/impl/api.py", line 350, in converted_call
#>     return _call_unconverted(f, args, kwargs, options, False)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/autograph/impl/api.py", line 478, in _call_unconverted
#>     return f(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/metrics.py", line 620, in update_state
#>     matches, sample_weight=sample_weight)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/keras/metrics.py", line 355, in update_state
#>     values = math_ops.cast(values, self._dtype)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/util/dispatch.py", line 201, in wrapper
#>     return target(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/ops/math_ops.py", line 964, in cast
#>     x = ops.convert_to_tensor(x, name="x")
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/profiler/trace.py", line 163, in wrapped
#>     return func(*args, **kwargs)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/ops.py", line 1540, in convert_to_tensor
#>     ret = conversion_func(value, dtype=dtype, name=name, as_ref=as_ref)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/constant_op.py", line 339, in _constant_tensor_conversion_function
#>     return constant(v, dtype=dtype, name=name)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/constant_op.py", line 265, in constant
#>     allow_broadcast=True)
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/constant_op.py", line 283, in _constant_impl
#>     allow_broadcast=allow_broadcast))
#>   File "/Users/ntustison/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/tensorflow/python/framework/tensor_util.py", line 553, in make_tensor_proto
#>     "supported type." % (type(values), values))

print( model )

#> Model
#> Model: "model_41"
#> ________________________________________________________________________________
#> Layer (type)              Output Shape      Param #  Connected to               
#> ================================================================================
#> input_4 (InputLayer)      [(None, 64, 64, 6 0                                   
#> ________________________________________________________________________________
#> conv3d_207 (Conv3D)       (None, 64, 64, 64 896      input_4[0][0]              
#> ________________________________________________________________________________
#> re_lu_76 (ReLU)           (None, 64, 64, 64 0        conv3d_207[0][0]           
#> ________________________________________________________________________________
#> conv3d_208 (Conv3D)       (None, 64, 64, 64 27680    re_lu_76[0][0]             
#> ________________________________________________________________________________
#> re_lu_77 (ReLU)           (None, 64, 64, 64 0        conv3d_208[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_10 (MaxPool (None, 32, 32, 32 0        re_lu_77[0][0]             
#> ________________________________________________________________________________
#> conv3d_209 (Conv3D)       (None, 32, 32, 32 55360    max_pooling3d_10[0][0]     
#> ________________________________________________________________________________
#> re_lu_78 (ReLU)           (None, 32, 32, 32 0        conv3d_209[0][0]           
#> ________________________________________________________________________________
#> conv3d_210 (Conv3D)       (None, 32, 32, 32 110656   re_lu_78[0][0]             
#> ________________________________________________________________________________
#> re_lu_79 (ReLU)           (None, 32, 32, 32 0        conv3d_210[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_11 (MaxPool (None, 16, 16, 16 0        re_lu_79[0][0]             
#> ________________________________________________________________________________
#> conv3d_211 (Conv3D)       (None, 16, 16, 16 221312   max_pooling3d_11[0][0]     
#> ________________________________________________________________________________
#> re_lu_80 (ReLU)           (None, 16, 16, 16 0        conv3d_211[0][0]           
#> ________________________________________________________________________________
#> conv3d_212 (Conv3D)       (None, 16, 16, 16 442496   re_lu_80[0][0]             
#> ________________________________________________________________________________
#> re_lu_81 (ReLU)           (None, 16, 16, 16 0        conv3d_212[0][0]           
#> ________________________________________________________________________________
#> max_pooling3d_12 (MaxPool (None, 8, 8, 8, 1 0        re_lu_81[0][0]             
#> ________________________________________________________________________________
#> conv3d_213 (Conv3D)       (None, 8, 8, 8, 2 884992   max_pooling3d_12[0][0]     
#> ________________________________________________________________________________
#> re_lu_82 (ReLU)           (None, 8, 8, 8, 2 0        conv3d_213[0][0]           
#> ________________________________________________________________________________
#> conv3d_214 (Conv3D)       (None, 8, 8, 8, 2 1769728  re_lu_82[0][0]             
#> ________________________________________________________________________________
#> re_lu_83 (ReLU)           (None, 8, 8, 8, 2 0        conv3d_214[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_36 (Conv (None, 8, 8, 8, 1 262272   re_lu_83[0][0]             
#> ________________________________________________________________________________
#> up_sampling3d_12 (UpSampl (None, 16, 16, 16 0        conv3d_transpose_36[0][0]  
#> ________________________________________________________________________________
#> concatenate_195 (Concaten (None, 16, 16, 16 0        up_sampling3d_12[0][0]     
#>                                                      re_lu_81[0][0]             
#> ________________________________________________________________________________
#> conv3d_215 (Conv3D)       (None, 16, 16, 16 884864   concatenate_195[0][0]      
#> ________________________________________________________________________________
#> re_lu_84 (ReLU)           (None, 16, 16, 16 0        conv3d_215[0][0]           
#> ________________________________________________________________________________
#> conv3d_216 (Conv3D)       (None, 16, 16, 16 442496   re_lu_84[0][0]             
#> ________________________________________________________________________________
#> re_lu_85 (ReLU)           (None, 16, 16, 16 0        conv3d_216[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_37 (Conv (None, 16, 16, 16 65600    re_lu_85[0][0]             
#> ________________________________________________________________________________
#> up_sampling3d_13 (UpSampl (None, 32, 32, 32 0        conv3d_transpose_37[0][0]  
#> ________________________________________________________________________________
#> concatenate_196 (Concaten (None, 32, 32, 32 0        up_sampling3d_13[0][0]     
#>                                                      re_lu_79[0][0]             
#> ________________________________________________________________________________
#> conv3d_217 (Conv3D)       (None, 32, 32, 32 221248   concatenate_196[0][0]      
#> ________________________________________________________________________________
#> re_lu_86 (ReLU)           (None, 32, 32, 32 0        conv3d_217[0][0]           
#> ________________________________________________________________________________
#> conv3d_218 (Conv3D)       (None, 32, 32, 32 110656   re_lu_86[0][0]             
#> ________________________________________________________________________________
#> re_lu_87 (ReLU)           (None, 32, 32, 32 0        conv3d_218[0][0]           
#> ________________________________________________________________________________
#> conv3d_transpose_38 (Conv (None, 32, 32, 32 16416    re_lu_87[0][0]             
#> ________________________________________________________________________________
#> up_sampling3d_14 (UpSampl (None, 64, 64, 64 0        conv3d_transpose_38[0][0]  
#> ________________________________________________________________________________
#> concatenate_197 (Concaten (None, 64, 64, 64 0        up_sampling3d_14[0][0]     
#>                                                      re_lu_77[0][0]             
#> ________________________________________________________________________________
#> conv3d_219 (Conv3D)       (None, 64, 64, 64 55328    concatenate_197[0][0]      
#> ________________________________________________________________________________
#> re_lu_88 (ReLU)           (None, 64, 64, 64 0        conv3d_219[0][0]           
#> ________________________________________________________________________________
#> conv3d_220 (Conv3D)       (None, 64, 64, 64 27680    re_lu_88[0][0]             
#> ________________________________________________________________________________
#> re_lu_89 (ReLU)           (None, 64, 64, 64 0        conv3d_220[0][0]           
#> ________________________________________________________________________________
#> conv3d_221 (Conv3D)       (None, 64, 64, 64 66       re_lu_89[0][0]             
#> ================================================================================
#> Total params: 5,599,746
#> Trainable params: 5,599,746
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> 
#>