Basic preprocessing pipeline for T1-weighted brain MRI

Various preprocessing steps that have been previously described in various papers including the cortical thickness pipeline:

preprocessBrainImage(
  image,
  truncateIntensity = c(0.01, 0.99),
  doBrainExtraction = TRUE,
  templateTransformType = NULL,
  template = "biobank",
  doBiasCorrection = TRUE,
  returnBiasField = FALSE,
  doDenoising = TRUE,
  intensityMatchingType = NULL,
  referenceImage = NULL,
  intensityNormalizationType = NULL,
  outputDirectory = NULL,
  verbose = TRUE
)

Arguments

image
input t1-weighted brain MRI

truncateIntensity
2-element vector giving the low and high quantiles for intensity truncation.

doBrainExtraction
boolean for performing ANTsRNet brain extraction If TRUE, returns mask. 3-D input only.

templateTransformType
see Details in help for antsRegistration. Typically "Rigid" or "Affine".

template
an ANTs image (not skull-stripped). Other premade templates are "biobank" and "croppedMni152".

doBiasCorrection
boolean for performing N4 bias field correction.

returnBiasField
if TRUE, return bias field as an additional output without bias correcting the preprocessed image.

doDenoising
boolean for performing non-local means denoising.

intensityMatchingType
Either "regression" or "histogram". Only is performed if !is.null(referenceImage).

referenceImage
reference image for intensity matching.

intensityNormalizationType
Either rescale the intensities to 0,1 (i.e., "01") or zero-mean, unit variance (i.e., "0mean"). If NULL normalization is not performed.

outputDirectory
destination directory for storing the downloaded template and model weights. Since these can be resused, if is.null(outputDirectory), these data will be downloaded to the subdirectory ~/.keras/ANTsXNet/.

verbose
print progress to the screen.

Value

preprocessed image and, optionally, the brain mask, bias field, and template transforms.

Details

\url{https://www.ncbi.nlm.nih.gov/pubmed/24879923}

Examples

library( ANTsR )
library( ANTsRNet )

image <- antsImageRead( getANTsRData( "r16" ) )
preprocessedImage <- preprocessBrainImage( image,
  truncateIntensity = c( 0.01, 0.99 ), doBrainExtraction = FALSE,
  doBiasCorrection = TRUE, doDenoising = TRUE,
  intensityNormalizationType = "01", verbose = FALSE )