Randomly transform image data (optional: with corresponding segmentations).

Apply rigid, affine and/or deformable maps to an input set of training images. The reference image domain defines the space in which this happens.

randomlyTransformImageData(referenceImage, inputImageList,
  segmentationImageList = NA, numberOfSimulations = 10,
  transformType = "affine", inputImageInterpolator = "linear",
  segmentationImageInterpolator = "nearestNeighbor", sdAffine = 1,
  numberOfControlPoints = 100, spatialSmoothing = 3, sdNoise = 1)

Arguments

referenceImage
defines the spatial domain for all output images. If the input images do not match the spatial domain of the reference image, we internally resample the target to the reference image. This could have unexpected consequences. Resampling to the reference domain is performed by testing using antsImagePhysicalSpaceConsistency then calling resampleImageToTarget upon failure.

inputImageList
list of lists of input images to warp. The internal list sets contains one or more images (per subject) which are assumed to be mutually aligned. The outer list contains multiple subject lists which are randomly sampled to produce output image list.

segmentationImageList
list of segmentation images corresponding to the input image list (optional).

numberOfSimulations
number of output images. Default = 10.

transformType
one of the following options c( "translation", "rigid", "scaleShear", "affine"," deformation" , "affineAndDeformation" ). Default = "affine".

inputImageInterpolator
one of the following options c( "linear", "gaussian", "bspline" ). Default = "linear".

segmentationImageInterpolator
one of the following options c( "nearestNeighbor", "genericLabel" ). Default = "nearestNeighbor".

sdAffine
parameter dictating deviation amount from identity for random linear transformations. Default = 1.0.

numberOfControlPoints
number of control points for simulated deformations. Default = 100.

spatialSmoothing
amount of spatial smoothing for simulated deformations. Default = 3.0.

sdNoise
randomization parameter for simulated deformations. Default = 1.0.

Value

list (if no directory set) or boolean for success, failure

Examples

library( ANTsR )
image1 <- antsImageRead( getANTsRData( "r16" ) )
image2 <- antsImageRead( getANTsRData( "r64" ) )
segmentation1 <- thresholdImage( image1, "Otsu", 3 )
segmentation2 <- thresholdImage( image2, "Otsu", 3 )
data <- randomlyTransformImageData( image1,
  list( list( image1 ), list( image2 ) ),
  list( segmentation1, segmentation2 ) )